% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_list.R
\name{add_level_to_factor}
\alias{add_level_to_factor}
\title{Merge factor vectors}
\usage{
add_level_to_factor(list_vec)
}
\arguments{
\item{list_vec}{a list of vectors}
}
\description{
Merge factor vectors
}
