% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DR.R
\name{add_min_max}
\alias{add_min_max}
\title{Append minimum and maximum values}
\usage{
add_min_max(test_data, ref_data)
}
\description{
Append the minimum and maximum values of \code{ref_data} into
\code{test_data}. This is to ensure that the same scaling to the \code{ref_data} will
be applied to \code{test_data} in subsequent computation
}
