% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{all_pathways_heat_pca}
\alias{all_pathways_heat_pca}
\title{Draw the heatmap and PCA on genes of different pathways}
\usage{
all_pathways_heat_pca(
  x,
  save_dir,
  all_path = NULL,
  highlight_cells = NULL,
  DE_dir = NULL,
  DE_label = NULL
)
}
\arguments{
\item{x}{a Seurat object}

\item{save_dir}{where outputs will be saved. For each pathway, one pdf file
will be made, including the PCA in the first page, heatmap the second, and
truncated heatmaps if the heatmap is too dense}

\item{all_path}{a dataframe with 2 columns: 'pathway' for the common names
of the pathway that is used for naming the output file, and 'kegg_id' for
the KEGG ID for that pathway. This ID may be obtained from:
https://www.genome.jp/kegg/pathway.html}

\item{highlight_cells}{which cells to highlight in the PCA and heatmap}
}
\description{
Draw the heatmap and PCA on genes of different pathways
}
