% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_DE.R
\name{all_vivo_all_vitro}
\alias{all_vivo_all_vitro}
\title{Compare 2 DE lists between all possible combination of cells}
\usage{
all_vivo_all_vitro(vitro_mark, vivo_mark, vitro_list, vivo_list)
}
\arguments{
\item{vitro_mark}{DE gene table between all possible pairwise comparisons in
the in vitro cell types}

\item{vivo_mark}{DE gene table between all possible pairwise comparisons in
the in vivo cell types}

\item{vitro_list}{all in vitro cells to be compared with each other}

\item{vivo_list}{all in vivo cells to be compared with each other}
}
\description{
Compare 2 DE lists between all possible combination of cells
}
