% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_preprocess.R
\name{amnioid_hist}
\alias{amnioid_hist}
\title{Plot distribution of features in amnioid}
\usage{
amnioid_hist(plot_df, xaxis, bin_group = 20, convert_radian = F)
}
\description{
Plot distribution of features in amnioid
}
