% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_abstract.R
\name{append_break_labels}
\alias{append_break_labels}
\title{Append the time interval labels}
\usage{
append_break_labels(
  vec,
  time_df,
  time_col,
  color_col,
  save_dir = NULL,
  save_label = NA,
  ...
)
}
\arguments{
\item{vec}{a vector to be labelled}

\item{time_df}{input dataframe to \code{break_time_line}}

\item{...}{arguments to pass to \code{break_time_line}}
}
\description{
This function uses the interval start and end points generated
from \code{break_time_line} to assign a time label for a vector
}
