% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_list.R
\name{append_if_numeric}
\alias{append_if_numeric}
\title{Append string before a numeric vector}
\usage{
append_if_numeric(x, append_str = "D")
}
\arguments{
\item{x}{string B}

\item{append_str}{string A}
}
\description{
Append a string A to the beginning of another string B if
string B can be coerced into numeric
}
