% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_poster.R
\name{arrange_poster}
\alias{arrange_poster}
\title{Put text and figures into a poster}
\usage{
arrange_poster(
  grob_list,
  save_path,
  config_arr,
  plot_width = 7,
  plot_height = 7,
  margin_width = 0.79,
  margin_height = 0.79,
  page_width = NULL,
  page_height = NULL
)
}
\arguments{
\item{grob_list}{a named list of objects to plot}

\item{save_path}{where to save the poster}

\item{config_arr}{a data.frame with the following columns:
\code{label}: label for the item that matches the name of the item in \code{grob_list}
\code{row.order}: order of the item along the row
\code{row.num}: number of rows
\code{column.start}: from which column the item span starts
\code{column.num}: number of columns
\code{text.style}: a string specifying the style of the text, available options
are 'fontsize', 'font_fam', 'fontface', 'xpos', 'ypos', 'hjust', 'vjust' and
any arguments that pass into \code{grid::grid.text}

The function calculate internally:
\code{column.end}: from which column the item span ends
\code{row.start}: from which row the item span starts
\code{row.end}: from which row the item span ends}
}
\value{
a saved pdf file
}
\description{
Empirally, I find one figure best correspond to 12 lines
}
