% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ax2D.R
\name{arrow_to_graph}
\alias{arrow_to_graph}
\title{Create ggplot layers from data for plotting arrow}
\usage{
arrow_to_graph(Arr, aes_param, nudge_ratio)
}
\arguments{
\item{Arr}{a list generated from \code{get_arrow_df}}
}
\description{
Create ggplot layers from data for plotting arrow
}
