% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCHeat_split.R
\name{average_by_group}
\alias{average_by_group}
\title{Average the expression matrix of Seurat by category}
\usage{
average_by_group(
  x,
  group.by,
  color_row,
  select_cells = NULL,
  slot_data = "data",
  assay = "RNA"
)
}
\arguments{
\item{x}{a Seurat object}

\item{group.by}{a vector whose first item is the feature on which averaging
is performed. Only the first item will be used in averaging. If a second
item is supplied, then a metadata field will appear containing the most
frequent item in the second category among the first category.}

\item{select_cells}{on which subset of cells averaging is performed}
}
\value{
a Seurat object
}
\description{
Average the expression matrix of Seurat by category
}
