% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ax3D.R
\name{Ax3D}
\alias{Ax3D}
\title{Draw 3D Axes}
\usage{
Ax3D(
  mapping = aes(group = 1),
  data = NULL,
  geom = "path",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{Arguements passed on to layer.
These are often aesthetics, used to set an
aesthetic to a fixed value, like color = "red" or size = 3.}

\item{theta}{The azimuthal direction in degrees.}

\item{phi}{The colatitude in degrees.}
}
\description{
This function adds 3D axes to a ggplot2 plot.
}
