% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correct.R
\name{batch_correct}
\alias{batch_correct}
\title{Perform Seurat batch correction}
\usage{
batch_correct(
  data_list,
  k_filter = NULL,
  dims = 1:30,
  refer = NULL,
  num_anchor_genes = 2000
)
}
\arguments{
\item{data_list}{a list of Seurat objects}
}
\description{
Perform Seurat batch correction
}
