% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_abstract.R
\name{break_time_line}
\alias{break_time_line}
\title{Obtain break points along a time line}
\usage{
break_time_line(
  x,
  time_col,
  color_col,
  time_order = NULL,
  abs_min = NULL,
  abs_max = NULL
)
}
\arguments{
\item{x}{a dataframe of all items}

\item{time_col}{which column in \code{x} has the time information}

\item{color_col}{which column in \code{x} has the grouping information}

\item{time_order}{order of the grouping labels in \code{color_col}. If not
supplied, it is assumed to follow the order of the mean time of each roup in
\code{color_col}}

\item{abs_min}{starting point of the time line. If NULL, it is assumed to be
the minimum time in \code{x}}
}
\description{
This function finds the interval of different periods along a
time line. This is not trivial, because sometimes different periods overlap
with each other. For aesthetic reasons however, there is significant demand
to show the periods as being mutually exclusive (which I strongly dislike).
As a result, this function will first find the min and max of time periods
in each group. The min is taken as the starting point. If the ending point
is the min of the next time period. If the min of the next time period
precedes that of its previous one, the starting point of the next time
period is taken to be the max of the previous time period.
}
