% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_cycle}
\alias{cell_cycle}
\title{Genes involved in cell cycle}
\format{
a dataframe, containing 5 columns: G1/S, S, G2/M, M, M/G1
Each column contains genes characteristic of the cell cycle phase
}
\source{
\url{https://www.cell.com/cell/fulltext/S0092-8674\%2815\%2900549-8}
}
\usage{
cell_cycle
}
\description{
Genes involved in cell cycle
}
\keyword{datasets}
