% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{cell_heat}
\alias{cell_heat}
\title{Plot correlation matrix}
\usage{
cell_heat(
  all_cor,
  metadata,
  features,
  column_scale = T,
  row_scale = F,
  col_rotation = 0
)
}
\arguments{
\item{all_cor}{correlation matrix, with both colnames and rownmaes}

\item{metadata}{the metadata for the items in the correlation matrix, does
not need to be in the exact order or shape, but its rownames should match
the rownames and colnames of \code{all_cor}}

\item{feature}{cluster by which feature in the metadata}
}
\description{
Plot correlation matrix
}
