% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{cell_violin}
\alias{cell_violin}
\title{Violin plot of cell-cell similarity}
\usage{
cell_violin(
  all_cor,
  metadata,
  feature,
  column_scale = T,
  row_scale = F,
  num_col = 1,
  AP = list()
)
}
\arguments{
\item{all_cor}{correlation matrix, with both colnames and rownmaes}

\item{metadata}{the metadata for the items in the correlation matrix, does
not need to be in the exact order or shape, but its rownames should match
the rownames and colnames of \code{all_cor}}

\item{feature}{cluster by which feature in the metadata}

\item{column_scale}{zero the mean across columns, make standard deviation 1}

\item{num_col}{number of columns in which the facets of the subplots will be
arranged}

\item{AP}{aesthetic parameters, \code{cell_order} controls the orderring of
cells in the violin plot legend and other fields for plotting.}
}
\description{
Take the mean of a correlation matrix \code{all_cor} of the cell
types or features along its \emph{column}
}
