% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{change_expre_level}
\alias{change_expre_level}
\title{Add the mean expression level of a cell type to igraph}
\usage{
change_expre_level(
  markers,
  igraph_net,
  celltype,
  group.by = "group",
  assay = "RNA",
  slot_data = "data",
  expr_col = "avgExpr",
  gene_col = "feature"
)
}
\arguments{
\item{markers}{a dataframe generated by \code{find_DE_genes}}
}
\description{
Add the mean expression level of a cell type to igraph
}
