% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_terms.R
\name{clean_terms}
\alias{clean_terms}
\title{Remove terms from GO/KEGG analysis that are not related to development}
\usage{
clean_terms(
  xx,
  AP,
  attr_name = "compareClusterResult",
  term_column = "Description"
)
}
\arguments{
\item{xx}{a compareClusterResult object}
}
\description{
Remove terms from GO/KEGG analysis that are not related to development
}
