% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_preprocess.R
\name{colony_count}
\alias{colony_count}
\title{Count the number of colonies}
\usage{
colony_count(dat, control_condition = "GFP", cell_lab = NULL)
}
\arguments{
\item{dat}{a dataframe containing the following columns
\code{series}: area ID in a culture well
\code{Date}: image batch
\code{cluster}: the cluster ID assigned to each \code{series}
\code{condition}: each siRNA condition}
}
\value{
a dataframe with the following columns:
\code{series}, \code{Date}, \code{condition}
\code{n}: number of colonies
\code{rel_n}: colony number relative to the mean in the control condition
}
\description{
This function first applies the following filter: remove DBSCAN
outliers, area > 400, number of cells per colony outside the range of 6~300
Secondly, it counts the number of colonies that meet such criteria.
}
