% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_utils.R
\name{colors2labels}
\alias{colors2labels}
\title{Convert WGCNA generated names to cluster names}
\usage{
colors2labels(color_vec, prefix = "cluster")
}
\description{
Convert WGCNA generated names to cluster names
}
