% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overRep.R
\name{compare_cluster_enrichment}
\alias{compare_cluster_enrichment}
\title{Over-representation test of genes in each cluster}
\usage{
compare_cluster_enrichment(
  markers,
  GO_data,
  logFC_term = "logFC",
  log_FC_thres = 0.25,
  enrich_area = "KEGG"
)
}
\arguments{
\item{markers}{a dataframe of DE genes from Seurat; alternatively, a list of
DE genes may be used}

\item{GO_data}{result from go_data}

\item{enrich_area}{either KEGG or GO or reactome}
}
\description{
Over-representation test of genes in each cluster
}
