% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overRep.R
\name{compare_reactome}
\alias{compare_reactome}
\title{Compare reactome across multiple cell types}
\usage{
compare_reactome(gene_list, cutoff = 0.05)
}
\description{
The \code{compareCluster} function does not work for 'enrichPathway' in my
computer. I need to write a similar function.
}
