% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CT}
\alias{CT}
\title{Cell type collection}
\format{
a list of cell types
\describe{
\item {TB_lineage}{the cell types within the trophoblast lineage}
\item {non_emb_lineage}{the cell types not within the embryonic lineage}
\item {in_vitro_cells}{the in vitro cell types}
}
}
\usage{
CT
}
\description{
Cell type collection
}
\keyword{datasets}
