% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{custom_color}
\alias{custom_color}
\title{Customise color}
\usage{
custom_color(
  feature_vec,
  aes_param = list(color_vec = NULL, date_color_vec = NULL),
  remove_NA = T,
  regexp = NULL
)
}
\arguments{
\item{feature_vec}{a vector of factors or characters of the names to plot}

\item{aes_param}{aesthetic parameters. The most important of which is
\code{color_vec}, a named vector for the colors with the names corresponding to
those in \code{feature_vec}. This needs not to be an exhaustive list. Any names
in \code{feature_vec} but not in \code{color_vec} will be automatically colored.}

\item{remove_NA}{remove NA values from \code{feature_vec}, i.e., do not color the
NA values}

\item{regexp}{which parts of the string in \code{feature_vec} to remove, in order
for it to match \code{color_vec}. This setting is useful when you try to cope
with varying names with the same \code{color_vec}.}
}
\description{
Customise color
}
