% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{custom_net}
\alias{custom_net}
\title{Customise igraph objects using ggraph}
\usage{
custom_net(igraph_net, AP = NULL, limits = NA, ranges = c(0, 15))
}
\arguments{
\item{igraph_net}{an igraph object}
}
\description{
Customise igraph objects using ggraph
}
