% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{custom_net_diff_nets}
\alias{custom_net_diff_nets}
\title{Multiple network graphs}
\usage{
custom_net_diff_nets(x, gene_list, markers, size_thres = 0, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{size_thres}{filter out nodes with expression levels smaller than a
certain threshold}

\item{...}{argments to pass to \code{custom_net}}
}
\description{
Multiple network graphs
}
