% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{custom_net_from_seurat}
\alias{custom_net_from_seurat}
\title{Create an igraph object from seurat}
\usage{
custom_net_from_seurat(
  x,
  genes,
  markers,
  celltype,
  thres = 0.6,
  size_thres = 0,
  ...
)
}
\arguments{
\item{...}{arguments to \code{add_pseudotime_to_net}}
}
\description{
a convenient wrapper for \code{plot_WGCNA_net}, then
\code{add_pseudotime_to_net} and \code{change_expre_level} and \code{filter_expression}
}
