% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{custom_scale}
\alias{custom_scale}
\title{Extend the axis scale}
\usage{
custom_scale(vec, x_y = "y", extend_ratio_min = 0, extend_ratio_max = 0)
}
\arguments{
\item{vec}{a vector as either the x or y coordinates of the data}

\item{x_y}{whether to extend x or y axis}

\item{extend_ratio_min}{how much to extend along the negative dimension of
the axis}

\item{extend_ratio_max}{how much to extend along the positive dimension of
the axis}
}
\value{
\code{xlim} or \code{ylim}
}
\description{
Extend the axis scale
}
