% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_poster.R
\name{custom_text}
\alias{custom_text}
\title{Customise text appearance from \code{grid.text}}
\usage{
custom_text(text_str, fontface, fontsize, font_fam, xpos, ypos, ...)
}
\arguments{
\item{text_str}{a single string}

\item{xpos}{position in the text horizontally as a percentage of the width}

\item{...}{arguments for \code{grid::grid.text}}
}
\description{
Customise text appearance from \code{grid.text}
}
