% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_poster.R
\name{custom_text_str}
\alias{custom_text_str}
\title{Customise text using an argument string}
\usage{
custom_text_str(text_str, text_args)
}
\arguments{
\item{text_args}{a string of keyword argument separated by comma, e.g.
'fontsize=5, fontfam='arial''}
}
\description{
Customise text using an argument string
}
