% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_analysis.R
\name{cycle_over_time}
\alias{cycle_over_time}
\title{Plot cycle score over pseudotime}
\usage{
cycle_over_time(
  plot_data,
  color_by,
  time_col,
  sel_phase = c("M", "S"),
  num_col = 1,
  extend_max = 0.4,
  AP = NULL
)
}
\arguments{
\item{plot_data}{a dataframe with cell cycle scores and pseudotime
information}

\item{color_by}{which column to color}

\item{time_col}{where the pseudotime information is stored}

\item{sel_phase}{which cell cycle phases to display, can be M, M.G1, G1.S, S, G2.M}

\item{num_col}{in how many columns the subplots are arranged}
}
\description{
Plot cycle score over pseudotime
}
\examples{
exp_mat_cc <- as.matrix (GetAssayData (all_data, assay='RNA', slot='data'))
ans <- get_phase_score (exp_mat_cc)
plot_data <- cbind (ans, all_data@meta.data)
cycle_over_time (plot_data, time_col='MGP_PT')
}
