% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPLVM_utils.R
\name{data_timeline}
\alias{data_timeline}
\title{Sample size across the timeline of the dataset}
\usage{
data_timeline(dat, x_col, y_col, AP = NULL)
}
\arguments{
\item{dat}{a dataframe e.g. from the metadata of a Seurat object}

\item{x_col}{which column contains the timeline (x axis)}

\item{y_col}{which column contains the dataset (y axis)}
}
\description{
Sample size across the timeline of the dataset
}
