% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_mat.R
\name{date_to_gene}
\alias{date_to_gene}
\title{Convert date-like strings into their gene names}
\usage{
date_to_gene(x)
}
\description{
Convert gene names that have been accidently converted into the
date format in excel back to gene names
}
