% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_motif.R
\name{DE_gene_pairwise_list}
\alias{DE_gene_pairwise_list}
\title{Generate gene list for each pair of comparison}
\usage{
DE_gene_pairwise_list(
  markers,
  compare_list,
  logFC_col = "logFC",
  gene_col = "feature",
  logFC_thres = 0.25,
  group1 = "group",
  group2 = "compare_group"
)
}
\arguments{
\item{markers}{data.frame generated by \code{find_DE_genes}}

\item{compare_list}{a list, each element has a 2-element vector}
}
\description{
Generate gene list for each pair of comparison
}
