% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_gene.R
\name{DE_lineage_genes}
\alias{DE_lineage_genes}
\title{Merge DE genes and known lineage markers}
\usage{
DE_lineage_genes(
  x,
  directory,
  top_number = 3,
  label = "all",
  feature = "revised",
  lineage_markers = NULL
)
}
\description{
Merge DE genes and known lineage markers
}
