% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCHeat.R
\name{determine_color_gradient}
\alias{determine_color_gradient}
\title{Obtain color gradient for heatmap}
\usage{
determine_color_gradient(xx, quantile_val = 0, center_scale = T, AP = NULL)
}
\arguments{
\item{xx}{a matrix or dataframe that will be the input of the heatmap}

\item{quantile_val}{the dynamic range of the heatmap. For example, if it is
0.05, the heatmap gradient will go from 5\% to 95\% percentile of the values
in \code{xx}}

\item{center_scale}{whether to color the middle of the dynamic range as
white. Otherwise, 0 is colored white.}
}
\description{
Obtain color gradient for heatmap
}
