% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DR.R
\name{dim_3_to_2}
\alias{dim_3_to_2}
\title{Perform coordinate transform from 3D to 2D}
\usage{
dim_3_to_2(dat, theta, phi, axes_names = c("x", "y", "z"))
}
\description{
Perform coordinate transform from 3D to 2D
}
