% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad_prog.R
\name{dim_plot_cell_similarity}
\alias{dim_plot_cell_similarity}
\title{Plot cell-cell similarity in a dimensionality reduction plot}
\usage{
dim_plot_cell_similarity(
  x,
  cell_sim,
  group.by,
  DR = "pca",
  dims = c(1, 2),
  aes_param = NULL
)
}
\arguments{
\item{x}{a Seurat object}

\item{cell_sim}{cell similarity matrix of N X M, where N is the sample and M
is the cell type. M must be equal to the column number of \code{x}}

\item{group.by}{which feature to color}

\item{DR}{dimensionality reduction assays in \code{x}}

\item{dims}{which dimensions to plot along the x and y axes}

\item{aes_param}{aesthetic parameters passing to \code{theme_TB}}
}
\description{
Plot cell-cell similarity in a dimensionality reduction plot
}
