% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DR_3D.R
\name{dim_red_3D}
\alias{dim_red_3D}
\title{Produce 3D scatter plot using gg3D}
\usage{
dim_red_3D(
  plot_data,
  x,
  y,
  z,
  color_by,
  all_theta = 0,
  all_phi = 0,
  show_axes = F,
  show_arrow = T,
  show_label = T,
  label_col = NULL,
  num_col = NULL,
  axis_length = 0.2,
  lab_just = 0.05,
  vert_just = 0,
  hor_just = 0,
  further_repel = F,
  repel_force = 1,
  fontface = "bold",
  size_highlight = NULL,
  highlight_ratio = 1.5,
  seg_color = NA,
  dim_name = "GPLVM",
  dim_vjust = 1,
  AP = NULL
)
}
\arguments{
\item{plot_data}{a dataframe with all the plotting information}

\item{x}{the column containing x coordinates, similarly for \code{y} and \code{z}}

\item{all_theta}{the degree of azimuthal (horizontal) rotation}

\item{all_phi}{the degree of vertical rotation}

\item{show_axes}{whehther to show 3D axes}

\item{show_arrow}{whether to show 3D segments (arrows)}

\item{show_label}{whether to show text labels}

\item{label_col}{which column contains the label information}

\item{num_col}{number of columns for \code{facet_wrap}}

\item{axis_length}{the proportion that the arrow axis occupies on the entire
axis}

\item{lab_just}{adjust the labels of the arrow axes radially. You may supply
one value for all 3 axes, all 3 values for x, y and z axes respectively}

\item{vert_just}{adjust the labels of the arrow axes vertically}

\item{further_repel}{if TRUE, the labels would be repelled away from the
data points as much as possible}

\item{size_highlight}{a character, logical or numeric vector specifying
which cells to magnify in size}

\item{highlight_ratio}{how much larger the highlighted cells should be}

\item{AP}{aesthetic parameters controlling arrow appearance}

\item{force_repel}{extent of repulsion}
}
\description{
Produce 3D scatter plot using gg3D
}
\author{
Yutong Chen
}
