% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DR.R
\name{dim_red_3D_traj}
\alias{dim_red_3D_traj}
\title{Add trajectory line to 3D scatterplot}
\usage{
dim_red_3D_traj(
  plot_data,
  px,
  py,
  pz,
  pcolor,
  traj_data,
  tx,
  ty,
  tz,
  tcolor,
  traj_color = "black",
  all_theta = 0,
  all_phi = 0,
  AP = NULL,
  repel_force = 1,
  further_repel = T,
  magnify_text = 1,
  label_traj_text = NULL,
  ...
)
}
\arguments{
\item{plot_data}{the dataframe for dim red plots}

\item{px, }{py, pz the column names in \code{plot_data} that corresponds to x, y
and z axes}

\item{pcolor}{the column name in \code{plot_data} that provide the color for
points}

\item{traj_data}{the dataframe for the trajectory data}

\item{tx, }{ty, tz, tcolor similar to px, py, pz, pcolor}

\item{...}{pass to \code{dim_red_3D}}
}
\description{
Add trajectory line to 3D scatterplot
}
\references{
\url{http://htmlpreview.github.io/?https://github.com/AckerDWM/gg3D/blob/master/gg3D-vignette.html}
}
\author{
Yutong Chen
}
