% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DR_3D.R
\name{DimPlot_3D}
\alias{DimPlot_3D}
\title{3D version of DimPlot}
\usage{
DimPlot_3D(
  x,
  feature,
  DR = "pca",
  dims = c(1, 2, 3),
  assay = "RNA",
  slot_data = "data",
  ...
)
}
\arguments{
\item{x}{a Seurat object}

\item{feature}{on which feature the color scheme is applied}

\item{DR}{which dimensionality reduction to use}

\item{...}{additional features to pass to \code{dim_red_3D}, including
\code{show_axes}, 'all_theta' and \code{all_phi}}
}
\description{
3D version of DimPlot
}
\author{
Yutong Chen
}
