% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DR.R
\name{DimPlot_labels}
\alias{DimPlot_labels}
\title{Obtain label positions}
\usage{
DimPlot_labels(dat, x_coord, y_coord, color_by, further_repel = T)
}
\arguments{
\item{further_repel}{if TRUE, the labels would be repelled away from the
data points as much as possible}
}
\description{
Obtain label positions
}
