% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overRep.R
\name{display_cluster_enrichment}
\alias{display_cluster_enrichment}
\title{Visualise results from over-representation test}
\usage{
display_cluster_enrichment(
  xx,
  show_graph = "emap",
  feature_vec = NULL,
  default_theme = F,
  save_dir = NULL,
  enrich_area = "GO",
  show_num = NULL,
  clean_results = T,
  simplification = T,
  AP = NULL,
  sim_dict = NULL,
  append_default_dict = T,
  subset_cluster = NULL,
  organism_db = NULL,
  num_col = NULL,
  markers = NULL,
  ...
)
}
\arguments{
\item{xx}{result from \code{compareCluster} or just a dataframe}

\item{show_graph}{whether to show the terms in piechart format ('emap') or
dotplot ('dotplot') or ridgeplot ('ridgeplot')}

\item{feature_vec}{to manually assign color}

\item{default_theme}{whether to use ggplot default theme}

\item{enrich_area}{either 'GO', 'KEGG' or 'reactome'. Only important for
saving the figure}

\item{clean_results}{whether to remove unrelated and disease related terms}

\item{simplification}{whether to simplify the GO/KEGG/reactome terms}

\item{sim_dict}{a data frame with 2 columns: 'ori' for the original terms,
'sub' for the strings that will replace the original terms. The default is a
a limited data frame built into this package.}

\item{append_default_dict}{append default dictionary to simplify the terms}
}
\description{
need to set seed to ensure reproducibility
}
