% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrich_bar}
\alias{enrich_bar}
\title{Barplot for GSEA results}
\usage{
enrich_bar(
  plot_data,
  organism_db,
  show_num = 4,
  markers = NULL,
  AP = NULL,
  show_gene_labels = 3,
  simplification = T,
  sim_dict = NULL,
  append_default_dict = T,
  compare_group_col = "compare_group",
  compare_group_name = "others",
  ...
)
}
\arguments{
\item{plot_data}{a dataframe generated from \code{run_GSEA_all_types}}

\item{organism_db}{a gene database}

\item{show_num}{how many terms to show}

\item{markers}{a dataframe generated from \code{find_DE_genes} that contain the
log fold change of gene expression. This information is used to order the
appearance of genes in the plot labels.}

\item{AP}{aesthetic parameters for plotting}

\item{show_gene_labels}{how many genes to show after each term}

\item{simplification}{whether to simplify the GO/KEGG/Reactome terms}

\item{sim_dict}{a data frame with 2 columns: 'ori' for the original terms,
'sub' for the strings that will replace the original terms. The default is a
a limited data frame built into this package.}

\item{append_default_dict}{append default dictionary to simplify the terms}

\item{compare_group_col}{which column in \code{markers} contain the
comparison/reference group information.}

\item{compare_group_name}{name of the comparison group if such a column does
not exist.}

\item{...}{other parameters to pass onto \code{gg_enrich_bar}. They include:}

\item{shrink_ratio}{by how much the fontsize of the labels should decrease}

\item{band_ratio}{how much thicker the arrow should be compared to the
\code{arrow_thickness} field in \code{AP}.}

\item{label_shift_ratio}{by what percentage along the x axis to shift the
term labels}

\item{nudge_x}{by how much the labels for the arrow to shift away from the
arrow head.}
}
\description{
Barplot for GSEA results
}
