% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{fast_subset}
\alias{fast_subset}
\title{Subset a Seurat object}
\usage{
fast_subset(
  x,
  sel_columns = NULL,
  sel_rows = NULL,
  assay = "RNA",
  slot_data = "data"
)
}
\arguments{
\item{x}{a Seurat object}
}
\description{
This function tries to perform subsetting in a faster way by
subsetting only the assay to be used and the VariableFeatures and the
metadata. The dimred objects are not subsetted.
}
