% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{filter_expression}
\alias{filter_expression}
\title{Delete those nodes with a small expression levels}
\usage{
filter_expression(graph_ob, size_thres)
}
\arguments{
\item{graph_ob}{an igraph object}

\item{size_thres}{the threshold of expression below which nodes are removed}
}
\value{
an igraph object
}
\description{
This function removes nodes from graph completely. Thus the
underlying graph structure will change during ploting. If you wish to
preserve the graph structure, use \code{hide_node}
}
