% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_gene.R
\name{filter_genes}
\alias{filter_genes}
\title{filter out genes with low expressions}
\usage{
filter_genes(
  x,
  quantile_val,
  assay = "RNA",
  slot_data = "data",
  method = "quantile"
)
}
\arguments{
\item{x}{a Seurat object}

\item{quantile_val}{filter away genes whose expressions are below a certain
percentile}
}
\description{
I am not really sure the most efficient way to do this given
that Seurat object does not have rowData. Therefore this function is quite
slow.
}
