% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_gene.R
\name{find_DE_genes}
\alias{find_DE_genes}
\title{Find DE genes across a particular feature}
\usage{
find_DE_genes(
  x,
  directory = NULL,
  group.by = "seurat_clusters",
  filename = NULL,
  label = "all"
)
}
\arguments{
\item{x}{a Seurat object}

\item{directory}{where to save the computed markers}

\item{group.by}{a column in the meta data to group the cell types}
}
\description{
Find DE genes across a particular feature
}
