% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_gene.R
\name{find_DE_pairwise}
\alias{find_DE_pairwise}
\title{Find pairwise DE gene expression}
\usage{
find_DE_pairwise(x, group.by, assay = "RNA", slot_data = "data")
}
\arguments{
\item{x}{a Seurat object}

\item{group.by}{a column in the meta data to group the cell types}
}
\description{
Find pairwise DE gene expression
}
