% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_utils.R
\name{find_eigengene}
\alias{find_eigengene}
\title{Find eigengene of WGCNA modules}
\usage{
find_eigengene(x, save_dir, cluster_num = "all", select_type = "revised")
}
\arguments{
\item{x}{a Seurat object}

\item{save_dir}{where to store the genes of each WGCNA cluster}

\item{cluster_num}{on which cell cluster WGCNA is performed. 'all' means the
entire dataset is involved}

\item{select_type}{where the cluster information is stored.}
}
\description{
Calculate the WGCNA connectivity matrix, perform hierarchical
clustering, cut tree, and then eigengene values. Weight power is
automatically determined
}
