% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_point.R
\name{find_peak}
\alias{find_peak}
\title{find the location of the maximum value}
\usage{
find_peak(x, time_index = NULL, max_only = T)
}
\arguments{
\item{x}{a dataframe with time series information along the columns, and the
features along the rows. It must contain one row that has the time index.
Otherwise, the time index is initiated as from 1 to the sample number}

\item{max_only}{detect absolute peak, instead of distinguishing between
positive or negative peak}
}
\value{
a dataframe with the following columns:
\code{val}: magnitude of the peak relative to the trough
\code{feature}: genes
\code{change_sign}: whether it is a positive peak or negative peak
\code{peak_time}: when the peak occurs
}
\description{
find the location of the maximum value
}
