% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_point.R
\name{find_transition_from_seurat}
\alias{find_transition_from_seurat}
\title{Detect genes that show the peak changes in GP}
\usage{
find_transition_from_seurat(
  seurat_ob,
  assay = "RNA",
  slot_data = "counts",
  time_ind = "pseudotime"
)
}
\arguments{
\item{seurat_ob}{a Seurat object. One row must contain the information of
pseudotime for each cell, as in the \code{time_ind} argument.}
}
\description{
Detect genes that show the peak changes in GP
}
