% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_gene.R
\name{gene_gene_mat}
\alias{gene_gene_mat}
\title{Gene expression in two categories}
\usage{
gene_gene_mat(x, compare.by, group.by, assay = "RNA", slot_data = "data")
}
\arguments{
\item{x}{a Seurat object}

\item{compare.by}{which features to be compared}

\item{group.by}{perform the analysis separately for different groups of
cells}
}
\value{
a dataframe
}
\description{
Calculate the mean expression of each gene in two categories
}
