% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_terms.R
\name{gene_per_enriched_term}
\alias{gene_per_enriched_term}
\title{Obtain genes of each enriched terms}
\usage{
gene_per_enriched_term(
  xx,
  term,
  organism_db,
  cell_type = NULL,
  category_col = "category"
)
}
\description{
For the data frame generated from enrichment analysis, achieves
the same function as \code{gene_per_term}
}
