% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_terms.R
\name{gene_per_term}
\alias{gene_per_term}
\title{For the results from \code{compareClusterResult} object}
\usage{
gene_per_term(xx, term, category_col = "Description", return_val = F)
}
\arguments{
\item{xx}{a compareClusterResult object}

\item{term}{a keyword that matches the term of interest}

\item{category_col}{which column in x that has the terms}

\item{return_val}{whether to return the gene names}
}
\description{
This function will print all the GO/KEGG/Reactome terms that
match the requested keyword or phrases with their p values. It will
optionally return a list of all the genes corresponding to the matches terms
}
\examples{
KG <- modules::use ('KEGG_path.R')
library (GOSemSim)
d <- godata('org.Hs.eg.db', ont="BP")
kk <- KG$compare_cluster_enrichment (markers, d, enrich_area='KEGG')
path_val <- KG$gene_per_term (kk, 'JAK-STAT', return_val=T)
}
