% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{get_all_path_view}
\alias{get_all_path_view}
\title{Pathview}
\usage{
get_all_path_view(
  genes,
  markers,
  cell_type,
  organism_db,
  save_dir,
  pathway_db = NULL,
  cluster_col = "feature",
  organism_name = "human",
  ...
)
}
\arguments{
\item{markers}{a dataframe of gene expression. It can be generated from
\code{find_DE_genes}}

\item{cell_type}{which cell type to analyse}

\item{organism_db}{the database for the organism to be investigated}

\item{save_dir}{where to save the results}

\item{pathway_db}{a dataframe with 2 columns: 'pathway' for the common names
of the pathway that is used for naming the output file, and 'kegg_id' for
the KEGG ID for that pathway. This ID may be obtained from:
https://www.genome.jp/kegg/pathway.html}

\item{gene}{a vector of gene names}
}
\description{
Plot the average expression of a list of genes in one cell type
on pathview plots for the major pathways in embryological development.
}
