% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ax2D.R
\name{get_arrow}
\alias{get_arrow}
\title{Obtain arrow object}
\usage{
get_arrow(AP)
}
\arguments{
\item{AP}{aesthetic parameter determining arrow types, relevant ones are
'arrow_angle', 'arrow_length', 'arrow_type'}
}
\value{
a ggplot arrow object
}
\description{
Obtain arrow object
}
