% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad_prog.R
\name{get_cell_similarity}
\alias{get_cell_similarity}
\title{Calculate cell-cell similarity using quadratic programming from DeConRNASeq}
\usage{
get_cell_similarity(
  x,
  group.by,
  select_types,
  compare_types = NULL,
  slot_name = "data",
  assay_name = "RNA"
)
}
\arguments{
\item{select_types}{for which cell types or cell features are similarity
estimates computed}

\item{compare_types}{to which cell types or cell features are those in
\code{select_types} compared with. If NULL, it would be every types except
\code{select_types}}
}
\description{
Calculate cell-cell similarity using quadratic programming from DeConRNASeq
}
