% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergence.R
\name{get_DE_from_KL}
\alias{get_DE_from_KL}
\title{Differential branching gene trajectoy analysis}
\usage{
get_DE_from_KL(x, num1, num2, divergence = "WD")
}
\arguments{
\item{x}{a dataframe produced by my BRGP script, containing the columns:
'branch' for branch information, 'x' for pseudotime, 'mean_' and 'var_' for
the mean and variance of each gene}

\item{num1}{the name of one of the branches to compare}

\item{num2}{similar to \code{num1}}

\item{divergence}{which divergence method to quantify the difference between
the 2 distributions. Only support 'WD' (2-Wasserstein distance) and 'KL'
(Kullback-Leibler divergence)}
}
\description{
Differential branching gene trajectoy analysis
}
