% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{get_gene_list}
\alias{get_gene_list}
\title{Obtain gene list from DE gene results}
\usage{
get_gene_list(
  markers,
  cluster_num,
  organism_db,
  gene_col = "feature",
  cluster_col = "group",
  FC = "logFC"
)
}
\description{
This function obtains a list of genes, the input format to GSEA
or over-representation test, from the top DE genes. For parameter setting,
see \code{run_GSEA}
}
