% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPLVM_utils.R
\name{get_gene_names}
\alias{get_gene_names}
\title{Remove the \code{mean_} and \code{var_} prefixes from a vector}
\usage{
get_gene_names(x)
}
\description{
Remove the \code{mean_} and \code{var_} prefixes from a vector
}
