% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{get_module_score}
\alias{get_module_score}
\title{Calculate the module score of KEGG pathway}
\usage{
get_module_score(
  x,
  save_path,
  all_path = NULL,
  pgenes = NULL,
  append_meta = F,
  regexp = "^X",
  species = "human",
  control_num = 50
)
}
\arguments{
\item{x}{a Seurat object}

\item{save_path}{where to save to output dataframe. It is recommended to
save the results because this function tends to be time consuming.}

\item{all_path}{a dataframe with \code{pathway} column being the comon pathway
names, and the \code{kegg_id} column being the KEGG ID for the pathways}

\item{pgenes}{a named list of vectors of genes}

\item{append_meta}{if TRUE, a Seurat object will be created with the
expression matrix being the module scores and metadata inherited from the
input Seurat object. This output will be more convenient for doing violin
plot and heatmap}

\item{regexp}{remove a part of cell ID names that may arise because of some
tidyverse manipulations. If NULL, no changes will happen}
}
\value{
a matrix or a Seurat object whose rows are gene module names in
\code{pgenes} and whose columns are cell names
}
\description{
It is extremely important to set seed for this function
}
