% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{get_path}
\alias{get_path}
\title{Obtain the path of dataset}
\usage{
get_path(root_dir, study)
}
\arguments{
\item{root_dir}{where the project is located. Under this root directory,
there must be two other directories called 'data' and 'results'}

\item{study}{name of the dataset to be loaded.}
}
\description{
This function returns a list of directores with \code{root} meaning
data directory and \code{save_dir} result directory
}
