% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{get_path_genes}
\alias{get_path_genes}
\title{Obtain the gene in a particular pathway}
\usage{
get_path_genes(pathway, path_name)
}
\arguments{
\item{pathway}{a dataframe with \code{pathway} column being the comon pathway
names, and the \code{kegg_id} column being the KEGG ID for the pathways}

\item{path_name}{the common name of a pathway}
}
\description{
Obtain the gene in a particular pathway
}
