% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_analysis.R
\name{get_phase_score}
\alias{get_phase_score}
\title{Assign phase score}
\usage{
get_phase_score(
  x,
  gene_vec = NULL,
  norm_factor = F,
  norm_column = F,
  norm_row = F
)
}
\arguments{
\item{x}{an expression matrix}

\item{gene_vec}{a vector of genes with their names being which cell cycles
they are expressed}
}
\description{
To calculate the phase score, Macosko (2015) used mean
expression of genes that mark a particular phase of the cell cycle. Not all
genes in the list are used. Only those that correlate with the mean
expression are included, followed by some normalisation. I am not certain
whether to apply normalisation or not, given that this research uses batch
combined data that has already been normalised. Qualitatively in a heatmap,
the results did not look much different.
}
