% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_abstract.R
\name{get_time_axis}
\alias{get_time_axis}
\title{Manually draw an x-axis}
\usage{
get_time_axis(
  min_x,
  max_x,
  x_name,
  AP,
  y_level = 0,
  shorten_ratio = 0.85,
  ticks_df = NULL,
  tick_height = 0.005
)
}
\arguments{
\item{min_x}{where to draw the x axis from}

\item{max_x}{where the axis ends}

\item{x_name}{name for the x axis}

\item{AP}{aesthetic paameters}

\item{y_level}{which y coordinate}

\item{shorten_ratio}{shorten the x axis from \code{max_x}. This is because
sometimes the axis tip is cut off from the edge of the plot.}

\item{ticks_df}{a dataframe for plotting the axis ticks}
}
\value{
a list of ggplot layers, must contain a column called \code{x} for the
locations of the ticks and the labels to put on
}
\description{
Sometimes you may want to put the x axis in certain locations
other than the periphery of the plot as in the default ggplot option. Here I
only implemented the version for x axis but it would be trivial for y axis
as well. The underlying rationale is simple. Draw the axis and put on the
labels.
}
