% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overRep.R
\name{gg_bar}
\alias{gg_bar}
\title{Plot barplot for overrepresented or enriched terms}
\usage{
gg_bar(
  xx,
  showCategory,
  organism_db = NULL,
  AP = NULL,
  rename_vec = NULL,
  show_gene_labels = 3,
  num_col = 3,
  markers = NULL
)
}
\arguments{
\item{xx}{a compareClusterResult subject}

\item{AP}{aesthetic parameters for the plot}
}
\description{
Plot barplot for overrepresented or enriched terms
}
