% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DR_2D.R
\name{gg_DimPlot}
\alias{gg_DimPlot}
\title{Reimplementation of DimPlot in Seurat for better graphic controls}
\usage{
gg_DimPlot(
  x,
  feature,
  DR = "pca",
  dims = c(1, 2),
  size_highlight = NULL,
  highlight_ratio = 1.5,
  label_col = NULL,
  further_repel = T,
  repel_force = 1,
  reverse_x = F,
  reverse_y = F,
  assay = "RNA",
  slot_data = "data",
  AP = NULL,
  seg_color = NA,
  plot_type = "dim_red",
  ...
)
}
\arguments{
\item{x}{a Seurat object}

\item{size_highlight}{a character, logical or numeric vector specifying
which cells to magnify in size}

\item{highlight_ratio}{how much larger the highlighted cells should be}

\item{label_col}{which column contains the label information}

\item{further_repel}{if TRUE, the labels would be repelled away from the
data points as much as possible}

\item{reverse_x}{reverse x axis direction. This is because sometimes most of
the contents are on the left, which would obstruct the arrow axis.}

\item{reverse_y}{similarly reverse y axis direction.}

\item{AP}{a list for \code{custom_color} and \code{theme_TB}}
}
\description{
Reimplementation of DimPlot in Seurat for better graphic controls
}
\author{
Yutong Chen
}
