% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DR.R
\name{gg_plot_dim_red}
\alias{gg_plot_dim_red}
\title{Return multiple subplots}
\usage{
gg_plot_dim_red(
  x,
  by_group,
  DR = "pca",
  dims = c(1, 2),
  highlight_font,
  size_highlight = NULL,
  return_sep = F,
  ...
)
}
\description{
Return multiple subplots
}
