% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_multi.R
\name{gg_silva}
\alias{gg_silva}
\title{Plot forest of GSEA terms}
\usage{
gg_silva(
  plot_df,
  AP,
  extend_x_neg = 1,
  extend_x_pos = 1,
  ladder_step = 1,
  band_width = 0.3,
  label_shift_ratio = 0.05,
  shrink_ratio = 1,
  nudge_y = 1,
  ...
)
}
\arguments{
\item{extend_x_neg}{how much percentage to extend along the left side of

the x axis to prevent the gene label text from being cut off.}

\item{extend_x_pos}{how much percentage to extend along the right side of
the x axis to prevent the gene label text from being cut off.}

\item{ladder_step}{how much should the two bars by separated from each
other. Empirically, I did not find this argument useful.}

\item{band_width}{the thickness of each bar}

\item{label_shift_ratio}{by what percentage along the x axis to shift the
term labels}

\item{shrink_ratio}{by how much the fontsize of the labels should decrease}

\item{nudge_y}{by how much the arrow and its labels to shift vertically up
from the maximum value along the y axis}

\item{...}{arguments for \code{sagittae}}

\item{band_ratio}{how much thicker the arrow should be compared to the
\code{arrow_thickness} field in \code{AP}.}
}
\description{
This function is the backend for \code{rich_forest}.
}
