% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GOsimp}
\alias{GOsimp}
\title{Simplify the GO/KEGG/Reactome terms}
\format{
a dataframe
\describe{
\item{ori}{the term to be replaced}
\item{sub}{the simplified term}
}
}
\usage{
GOsimp
}
\description{
Simplify the GO/KEGG/Reactome terms
}
\keyword{datasets}
