% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_abstract.R
\name{graph_abs_time}
\alias{graph_abs_time}
\title{Generate illustration plot for pseudotime progression}
\usage{
graph_abs_time(cell_df, time_df, AP = NULL, off_set = 8, thickness = 0.01, ...)
}
\arguments{
\item{cell_df}{a dataframe containing cell type information along the
pseudotime}

\item{time_df}{a dataframe containing names for pseudotime intervals}

\item{...}{other arguments for \code{plot_two_branch}}
}
\value{
a ggplot object
}
\description{
This function is very specific to this paper and is not useful
for general applications.
}
