% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_utils.R
\name{hub_score}
\alias{hub_score}
\title{Calculate intra modular connectivity}
\usage{
hub_score(x, weight_power, save_dir, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{...}{parameters for \code{subset_celltype}}
}
\description{
Calculate intra modular connectivity
}
