% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_list.R
\name{incorporate_TSC}
\alias{incorporate_TSC}
\title{Incorporate Trophoblast stem cells into Seurat object}
\usage{
incorporate_TSC(x, analyse_by = "new_cluster2", add_by = "revised")
}
\description{
a specific function unlikely to be helpful
}
