% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPLVM_utils.R
\name{integrated_seurat}
\alias{integrated_seurat}
\title{Integrate expression matrix, branch information, pseudotime data into a
single Seurat object.}
\usage{
integrated_seurat(exp_mat, pseudotime, branch, metadata)
}
\description{
I did not find this function helpful because later I added them
to the metadata of the integrated dataset.
}
