% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KeggID}
\alias{KeggID}
\title{KEGG ID conversion}
\format{
a dataframe
\describe{
\item{pathway}{the common name of a pathway}
\item{kegg_id}{the KEGG ID of that pathway}
}
}
\usage{
KeggID
}
\description{
KEGG ID conversion
}
\keyword{datasets}
