% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correct.R
\name{label_transfer}
\alias{label_transfer}
\title{Transfer the known labels from selected datasets to other datasets}
\usage{
label_transfer(merged_data, feature, by_paper = NULL)
}
\arguments{
\item{merged_data}{a Seurat object}

\item{feature}{a column in the metadata of the \code{merged_data}, which
identifies the label to be transferred}

\item{by_paper}{select which datasets from which paper are used as a
reference. The default is the dataset(s) that contain missing values in the
field specified by the \code{feature} argument}
}
\description{
This function uses Seurat \code{FindTransferAnchors} and
\code{TransferData} functions to perform label transfer
}
