% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lineage_markers}
\alias{lineage_markers}
\title{Lineage marker genes}
\format{
a named vector of genes, with the names being the lineages that they
mark
}
\usage{
lineage_markers
}
\description{
Lineage marker genes
}
\keyword{datasets}
