% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_list.R
\name{load_correct_label}
\alias{load_correct_label}
\title{Load the corrected label of cells}
\usage{
load_correct_label(
  metadata,
  root_dir,
  file_name,
  sheetName,
  barcode_col = NULL,
  process_ID_func = NULL
)
}
\arguments{
\item{metadata}{metadata to be added to seurat}

\item{root_dir}{directory to the correct cell labels}

\item{file_name}{file_name of the correct cell labes}

\item{sheetName}{the excel sheet name}

\item{barcode_col}{the column storing the barcode, default is the rownames}

\item{process_ID_func}{function to preprocess the cell barcode}
}
\description{
Load the corrected label of cells
}
