% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCHeat.R
\name{make_anno_legend}
\alias{make_anno_legend}
\title{Create annotation legend manually}
\usage{
make_anno_legend(anna_param, legend_bar_title, color, name = NA)
}
\arguments{
\item{anna_param}{annotation param for \code{annotation_legend_param}}

\item{legend_bar_title}{what to appear in the title of legend bars in the
order of appearance}

\item{color}{the color for those titles in \code{legend_bar_title}}

\item{name}{name of the legend}
}
\description{
This function takes in similar arguments as \code{HeatmapAnnotation}
and output the legend only. This is to give a better graphics control over
the legend position.
}
