% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_analysis.R
\name{make_cycle_heat}
\alias{make_cycle_heat}
\title{Make a heatmap for phase score}
\usage{
make_cycle_heat(x, seurat_ob, feature = "revised", AP = NULL)
}
\arguments{
\item{x}{a matrix containing the phase scores}
}
\description{
Make a heatmap for phase score
}
