% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_utils.R
\name{make_eigen_violin}
\alias{make_eigen_violin}
\title{Violin plot for eigengene value in each cell cluster}
\usage{
make_eigen_violin(x, datME, feature = "revised")
}
\description{
Violin plot for eigengene value in each cell cluster
}
