% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPLVM_utils.R
\name{mean_log_likelihood}
\alias{mean_log_likelihood}
\title{Calculate mean log likelihood}
\usage{
mean_log_likelihood(mu, variance, x)
}
\description{
This function works extremely slow. I recommend using the
python alternative.
}
