% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correct.R
\name{merge_seurat}
\alias{merge_seurat}
\title{Merge Seurat object}
\usage{
merge_seurat(list_obj, assays, slot_data = "data")
}
\arguments{
\item{list_obj}{a list of Seurat objects}

\item{assays}{a vector of assay names to be merged. If only one name is
supplied, this assay from all Seurat objects all be combined.}
}
\description{
Merge seurat assay data and meta.data from different Seurat
objects. The results will be stored in the slot 'data' and assay name 'RNA'.
Unlike the \code{merge} function provided by Seurat, this works much faster
because not all attributes in the Seurat objects are merged
}
