% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_list.R
\name{mixed_order_sort}
\alias{mixed_order_sort}
\title{Slight modification to \code{mixedsort}}
\usage{
mixed_order_sort(vec)
}
\description{
\code{gtools::mixedsort} is a wonderful function. However, if a
string contains a hyphen followed by a number, the function would mistake it
as a negative function and sort the values accordingly. This function
prevents \code{mixedsort} from doing so.
}
