% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_DE.R
\name{p_val_all_pairs}
\alias{p_val_all_pairs}
\title{Comparison p-value between 2 groups of DE genes}
\usage{
p_val_all_pairs(comp_list, vitro_list, vivo_list)
}
\arguments{
\item{comp_list}{output from \code{all_vivo_all_vitro }}

\item{vitro_list}{all in vitro cells to be compared with each other}

\item{vivo_list}{all in vivo cells to be compared with each other}
}
\value{
a dataframe with the following attributes:
\code{model}: the cell types in comparison. For example, 'bTSC_vs_hTSC-OKAE,
ICM_vs_TB' assumes that bTSC is similar to ICM/TB while hTSC-OKAE is similar to
the opposite TB/ICM.
\code{ratio}: a/c (see the Method section below)
\code{rev_ratio}: b/d
\code{pval}: Fisher's exact test p value
\code{padj}: adjusted p value for false discovery rate
}
\description{
First, select the true positive (TP) of the intersection between the following
permutations:\tabular{lll}{
   TP \tab vitro1 up \tab vitro2 up \cr
   vivo1 up \tab a \tab b \cr
   vivo2 up \tab c \tab d \cr
}


The null hypothesis is that the upregulated genes in vitro1 compared with
vitro2 (denoted as vitro1_vs_vitro2) do not intersect with vivo1_vs_vivo2 more
than vivo2_vs_vivo1, compared with vitro2_vs_vitro1.

For example, if vitro1 and vitro2 are very similar, then they should both
contain a similar number of vivo1_vs_vivo2, and/or vivo2_vs_vivo1. If vitro1
and vitro2 are different in the direction of vivo1 and vivo2, then vitro1
should contain a different vivo1 to vivo2 ratio, compared with vitro2.

The probability of evaluating this null hypothesis is evaluated by Fisher's
exact test:

$$ p = \frac {(a+b)! (c+d)! (a+c)! (b+d)! }{a! b! c! d! (a+b+c+d)!} $$
}
