% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_list.R
\name{partial_relevel}
\alias{partial_relevel}
\title{Partially relevel a vector}
\usage{
partial_relevel(
  x,
  reorder_levels = NULL,
  mixed_sorting = F,
  leading_char = "",
  skip_char = 0
)
}
\arguments{
\item{x}{a vector}

\item{reorder_levels}{a vector specifying which celltype should be in the
earlier positions of of the factor level. There is no need to specify a
string that matches the celltype name exactly. Regular expression will be
used to match the string names from the beginning of the strings.}

\item{mixed_sorting}{if set TRUE, the \code{mixedsort} function from gtools package
will be used and will override the \code{reorder_levels} argument}

\item{leading_char}{ignore the leading character when releveling}
}
\value{
a factor vector
}
\description{
Turn a vector into a factor vector. The order of the vector is
not changed, only the factor level
}
\examples{
library (gtools)
reorder_levels <- c('Zy', '4C', '8C') # 'Zy' would be shown first in  the
# factor level, '4C' would be the second etc
celltypes <- c( '4C_D2', '8C_D2', 'Zy_D1' , '4C_D2', 'cMor_D3')
partial_relevel ( celltypes, reorder_levels )
#alternatively you may use the default ordering vector
}
