% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{path_eigen}
\alias{path_eigen}
\title{Draw the heatmap for the eigengene for single cell}
\usage{
path_eigen(
  x,
  all_path = NULL,
  threshold = 0.02,
  select_cells = NULL,
  return_mat = F
)
}
\arguments{
\item{x}{a Seurat object}

\item{threshold}{This function assigns the cells with the highest total gene
expressions in a particular pathway to have positive eigen values. This
\code{threshold} argument sets the range of cells to calculate the total
expression. For example, a threshold of 0.02 means that the total expression
will be calculated for the bottom 2\% and top 2\% of the cells}
}
\description{
Due to the problem with the sign of eigenvector, this function
is not recommendd to visualise pathway activity. Instead, you may use the
\code{get_module_score} function and plot the results using \code{seurat_heat}
}
\examples{
path_eigen (all_data)
}
