% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{pathway_heat_pca}
\alias{pathway_heat_pca}
\title{Save the PCA and heatmap of genes in a pathway}
\usage{
pathway_heat_pca(
  x,
  gene_list,
  save_dir,
  cutoff = 70,
  highlight_cells = NULL,
  group.by = c("revised", "date"),
  DE_dir = NULL,
  DE_label = NULL
)
}
\arguments{
\item{cutoff}{If the \code{gene_list} argument is longer than a cutoff length,
then the \code{gene_list} will be truncated into lists of genes, each of which is
no longer than the cutoff length. Then each heatmap will be generated for
each sublist}
}
\description{
The results will be stored in a pdf file. For each pathway, a
PC plot using the genes in that pathway is shown, followed by expression of
genes in that pathway in heatmap format.
}
