% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPLVM_utils.R
\name{pca_with_pt_line}
\alias{pca_with_pt_line}
\title{PCA plot of cells with pseudotime trajectory projected onto it}
\usage{
pca_with_pt_line(
  pc_pt,
  pt_mat,
  metadata,
  color.by,
  branch_info = NULL,
  proj_var = NULL,
  num_dim = c(1, 2),
  AP = AP
)
}
\arguments{
\item{pc_pt}{a prcomp object}

\item{pt_mat}{the matrix for pseudotime trajectory}

\item{branch_info}{a vector for branch assignment}
}
\description{
The reason why I have not integrated PCA computation into this
function is that pca takes too long for large matrices. That's why the
function looks quite inconvenient. I use \code{gmodels::fast.prcomp} for PCA.
Similarly, users need to calculate the PC projected variance externally e.g.
proj_var <- solve (pc_pt$rotation^2, t(pred_list[\link{2}]), tol=1e-20)

Strictly speaking, projecting variance onto PCs require NNLS. I have
implemented it in python but no time to do so in R. Fortunately, I did not
need to use this function often.
}
