% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_point.R
\name{peak_gene_for_heatmap}
\alias{peak_gene_for_heatmap}
\title{Get genes for labelling in heatmap}
\usage{
peak_gene_for_heatmap(xx, plot_quantile = 0.95, min_genes = 2)
}
\arguments{
\item{xx}{a dataframe generated from \code{find_peak}. For the columns that it
should contain, see \code{find_peak}}

\item{plot_quantile}{the top quantile of genes in \code{xx} to show in heatmap}

\item{min_genes}{minimum number of genes to show in the heatmap row labels}
}
\description{
Get genes for labelling in heatmap
}
