% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_analysis.R
\name{phase_density_plot}
\alias{phase_density_plot}
\title{Plot the distribution of cycle phase scores as ridge/wave like plots}
\usage{
phase_density_plot(
  ans,
  x,
  corr_feature,
  select_phase = NULL,
  num_row = 2,
  AP = NULL
)
}
\description{
same usage as \code{phase_PC_plot}
}
