% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_analysis.R
\name{phase_PC_plot}
\alias{phase_PC_plot}
\title{Correlation of phase score with PC}
\usage{
phase_PC_plot(ans, x, corr_feature, DR = "pca", num_dim = 1, AP = NULL)
}
\arguments{
\item{x}{a seurat object}

\item{corr_feature}{which feature is the selected DR correlated with}

\item{num_dim}{which dimension in the dimensionality reduction (DR) method to use}
}
\description{
Correlation of phase score with PC
}
\examples{
exp_mat <- as.matrix (Seurat::GetAssayData (all_data, assay='RNA', slot='data'))
ans <- get_phase_score (exp_mat )
phase_PC_plot (ans, all_data, 'broad_type')
}
