% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_all_WGCNA_nets}
\alias{plot_all_WGCNA_nets}
\title{Plot the gene-gene interaction network for all WGCNA clusters}
\usage{
plot_all_WGCNA_nets(x, all_groups, save_dir, ...)
}
\arguments{
\item{x}{a Seurat object}

\item{save_dir}{where to save the plot}

\item{color_row}{a list of vector of gene names}
}
\description{
This is basically repeating the \code{plot_WGCNA_net} function and
save every plot in a single pdf file.
}
