% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correct.R
\name{plot_BC_filter}
\alias{plot_BC_filter}
\title{Investigate whether varying the size of k.filter can affect the
result of batch correction}
\usage{
plot_BC_filter(data_list, directory, start_from = 10)
}
\arguments{
\item{data_list}{list of Seurat objects to be batch corrected}

\item{directory}{where the results of the batch corrections will be
saved}

\item{start_from}{the starting point of testing k.filter size. The
end point is 200 by default.}
}
\description{
Investigate whether varying the size of k.filter can affect the
result of batch correction
}
