% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_preprocess.R
\name{plot_colony_count}
\alias{plot_colony_count}
\title{Boxplot of the colony features in each siRNA condition}
\usage{
plot_colony_count(
  dat,
  y_val = "rel_num_colony",
  shape_by = "Date",
  control_condition = "GFP",
  plot_pval = "p.signif",
  plot_type = "box",
  paired_test = F,
  display_shape = T,
  new_level = NULL,
  AP = NULL
)
}
\arguments{
\item{dat}{a dataframe from \code{colony_count}}

\item{y_val}{y axis of the boxplot}

\item{control_condition}{which is the control group in the \code{condition}
column}

\item{plot_pval}{which pvalue format to plot, can be 'p.adj' for the BH
adjusted p value score, 'p.signif' for the symbol for 'p.adj' or 'p' for
unadjusted p value}
}
\description{
Boxplot of the colony features in each siRNA condition
}
