% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_gene.R
\name{plot_DE_genes}
\alias{plot_DE_genes}
\title{Visualise DE genes}
\usage{
plot_DE_genes(
  x,
  markers,
  top_gene = 1,
  by_group = "cluster",
  show_group = "Type",
  show_heatmap = TRUE
)
}
\arguments{
\item{x}{Seurat object}

\item{markers}{DE gene marker dataframe computed by
\code{FindAllMarkers}. Alternatively, one can parse a vector of marker
genes.}

\item{top_gene}{how many genes to show per group}

\item{by_group}{which group of features in which the top DE genes
are selected}

\item{show_group}{which group of features in which the expression of
top DE genes are plotted against}

\item{show_heatmap}{whether to show heatmap or violin plot}
}
\description{
Visualise DE genes
}
