% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_red.R
\name{plot_dim_red}
\alias{plot_dim_red}
\title{Quick way of plotting DimPlot in Seurat}
\usage{
plot_dim_red(
  input_data,
  group.by,
  DR = "pca",
  select_cells = NULL,
  dims = c(1, 2),
  return_sep = F,
  num_col = NULL,
  AP = NULL,
  ...
)
}
\arguments{
\item{input_data}{a Seurat object}

\item{group.by}{a 2-element vector, parsed to the \code{group.by} argument of
\code{DimPlot} in Seurat}

\item{DR}{the type of dimensionality reduction method}

\item{select_cells}{cells to be shown in the plot}

\item{dims}{two axes of the reduced dimensional space}

\item{num_col}{how many columns the subplots are arranged}

\item{plot_3D}{whether to plot in 3D}
}
\value{
a ggplot or list of ggplot objects
}
\description{
Please run the \code{run_dim_red} function before executing this
function
}
