% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_red.R
\name{plot_gene_PC}
\alias{plot_gene_PC}
\title{Plot the contribution of genes to each PCs}
\usage{
plot_gene_PC(
  x,
  show_markers = NULL,
  dims = c(1, 2),
  color_by = "Type",
  directory = NULL,
  label = "all",
  DR = "pca",
  skip_char = 0,
  color_markers = NULL,
  DE_rank = "logFC",
  cluster_col = "group",
  feature_col = "feature",
  AP = NULL
)
}
\arguments{
\item{x}{a Seurat object}

\item{show_markers}{the genes to shown in the final plot. If \code{NULL}, all genes
will be shown. If \code{TF}, a list of transcriptional factors will be shown.
Alternatively, the argument accepts a vector of genes.}

\item{dims}{which pcs to plot}

\item{color_by}{which feature of genes to color, default is the cell
type in which the genes are differentially expressed}
}
\description{
Plot the contribution of genes to each PCs
}
