% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_abstract.R
\name{plot_one_branch}
\alias{plot_one_branch}
\title{Plot sigmoid curve for one branch}
\usage{
plot_one_branch(
  dat,
  branch_id,
  branch_col = "epil_branch",
  slope = 1,
  off_set = 5,
  filter_lim = c(-Inf, Inf),
  time_col = "MGP_PT",
  br_sign = 1,
  color_col = "broad_type",
  thickness = 0.01,
  select_cells = NULL,
  label_group = NULL,
  text_offset = 0,
  AP = NULL,
  elevation = 0,
  time_df = NULL,
  time_time_col = NULL,
  time_color_col = NULL,
  time_group_col = NULL,
  branch_name = NA,
  repel_force = 0.001,
  save_dir = NULL,
  plot_branch_names = T
)
}
\arguments{
\item{dat}{a dataframe}

\item{branch_id}{which branch(es) to plot}

\item{branch_col}{which column contains the branch information}

\item{slope}{the steepness of the curve. The larger the value, the smoother
the curve becomes}

\item{off_set}{shift the tuning point of the sigmoid curve. The higher the
value, the more leftwards the curve will shift}

\item{filter_lim}{the min and max of x axis}

\item{time_col}{which column contains pseudotime information}

\item{br_sign}{whether to flip the curve along the x axis, -1 indicates a
flip}

\item{color_col}{which column contains grouping information}

\item{thickness}{the thickness of the curve}

\item{select_cells}{which cells in \code{dat} to plot}

\item{text_offset}{how much the labels deviate from the top of the ribbon}

\item{elevation}{how much the curve is vertically translated. The direction
is the same as \code{br_sign}}
}
\value{
a list of ggplot layers
}
\description{
Plot sigmoid curve for one branch
}
