% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_abstract.R
\name{plot_two_branch}
\alias{plot_two_branch}
\title{Draw 2 sigmoid curves}
\usage{
plot_two_branch(
  dat,
  branch_col = "epil_branch",
  color_col = "broad_type",
  time_col = "MGP_PT",
  branch1 = c("main", "EVT_b"),
  branch2 = c("main", "STB_b"),
  select_cells1 = NULL,
  select_cells2 = NULL,
  AP = NULL,
  time_axis = T,
  axis_label = "time",
  ...
)
}
\arguments{
\item{branch1}{which samples in \code{dat} make up the first branch}

\item{select_cells1}{which cells in the first branch}

\item{time_axis}{whether to plot the x axis}

\item{axis_label}{what to label the \code{time_axis}}
}
\description{
Draw 2 sigmoid curves
}
\examples{
plot_two_branch (meta, 
                 select_cells1 = meta$broad_type != 'STB',
                 select_cells2 = meta$broad_type != 'EVT',
                 text_offset=0.007, off_set= 5
) + ggplot2::theme (legend.position='none')
}
\seealso{
\code{plot_one_branch}
}
