% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_WGCNA_net}
\alias{plot_WGCNA_net}
\title{Plot gene-gene correlation network in a WGCNA cluster}
\usage{
plot_WGCNA_net(
  x,
  color_row,
  celltype = "all",
  feature = "revised",
  scale_edge_weight = 0.2,
  scale_node_size = 0.2,
  threshold = 0.5,
  return_igraph = F
)
}
\arguments{
\item{x}{a Seurat object}

\item{color_row}{a vector of gene names}

\item{feature}{which metadata entry the cell type information is stored.
Ignore it if \code{cell_type='all'}}

\item{scale_edge_weight}{adjust the weight of the lines}

\item{scale_node_size}{adjust the relative node size}

\item{threshold}{only gene-gene pair with correlation beyond a threshold
will be selected}

\item{return_igraph}{whether igraph object will be returned or whether it
will be plotted}

\item{cell_type}{the cell type in which correlation is computed. By default,
all cell types will be used}
}
\description{
Plot gene-gene correlation network in a WGCNA cluster
}
