% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess the dataset generated from the quantification pipeline}
\usage{
preprocess(DAPI, bright, tree)
}
\arguments{
\item{DAPI}{a dataframe with information for each nucleus in the DAPI
channel}

\item{bright}{a dataframe with information for each organoid}

\item{tree}{a dataframe describing the relationship between amnioid and
nuclei, required columns: child, parent, condition}
}
\description{
Preprocess the dataset generated from the quantification pipeline
}
