% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPLVM_utils.R
\name{preprocess_df}
\alias{preprocess_df}
\title{Return parts of the dataframe \code{x_df} that begins with \code{sel_col} and with
rows being \code{genes}}
\usage{
preprocess_df(x_df, sel_col, genes)
}
\description{
Return parts of the dataframe \code{x_df} that begins with \code{sel_col} and with
rows being \code{genes}
}
