% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPLVM_utils.R
\name{pseudo_real_time}
\alias{pseudo_real_time}
\title{Plot pseudotimee against real time}
\usage{
pseudo_real_time(
  dat,
  real_time,
  pseudotime,
  color_by,
  AP = NULL,
  rotation = 90,
  ...
)
}
\arguments{
\item{dat}{a dataframe}

\item{real_time}{the column in \code{dat} that stores the real time information}

\item{pseudotime}{the column in \code{dat} that stores the pseudotime information}

\item{color_by}{the column that stores the features to colors}
}
\description{
Plot pseudotimee against real time
}
