% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPLVM_utils.R
\name{pseudotime_heat}
\alias{pseudotime_heat}
\title{Heatmap shows gene expression along the order of pseudotime across groups}
\usage{
pseudotime_heat(
  seurat_list,
  show_genes,
  group.by,
  order_row = "pseudotime",
  assay = "RNA",
  slot_data = "counts",
  return_sep = F,
  col_title = NULL,
  ...
)
}
\description{
I did not find this function particularly helpful.
\code{seurat_heat} is more flexible.
}
