% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_mat.R
\name{remove_ribosomal_genes}
\alias{remove_ribosomal_genes}
\title{Remove ribosomal genes from single cell dataset}
\usage{
remove_ribosomal_genes(x, row_is_gene = TRUE)
}
\arguments{
\item{x}{can be S3 or S4 class}
}
\description{
Remove ribosomal genes from single cell dataset
}
