% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DR.R
\name{rescaling}
\alias{rescaling}
\title{Scale data into a given range}
\usage{
rescaling(vec, to_range = c(0, 1))
}
\arguments{
\item{vec}{a numeric vector}

\item{to_range}{range of the output data}
}
\description{
Scale data into a given range
}
