% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_multi.R
\name{rich_forest}
\alias{rich_forest}
\title{Barplot for GSEA results}
\usage{
rich_forest(
  plot_data,
  markers,
  organism_db,
  show_num = 4,
  show_gene_labels = 3,
  category_col = "category",
  pval_col = "p.adjust",
  enrich_col = "value",
  term_col = "termID",
  group1 = "cluster",
  group2 = "compare_group",
  simplification = T,
  sim_dict = NULL,
  append_default_dict = T,
  compare_group_col_marker = "compare_group",
  AP = NULL,
  ...
)
}
\arguments{
\item{plot_data}{a dataframe generated from \code{run_GSEA_all_types}. NB: if the
input data frame is from other functions, make sure you apply some p value
cut off. It must have the following columns:}

\item{markers}{a dataframe generated from \code{find_DE_genes} that contain the
log fold change of gene expression. This information is used to order the
appearance of genes in the plot labels. It must contain the columns:
\code{group} (group1), \code{logFC} (used to rate which genes to show), \code{feature}
(genes), and \code{compare_group} (group2). You can change \code{compare_group} via:}

\item{organism_db}{a gene database}

\item{show_num}{how many terms to show per pairwise comparison}

\item{show_gene_labels}{how many genes to show after each term}

\item{category_col}{the GSEA terms}

\item{pval_col}{the (adjusted) p value}

\item{enrich_col}{the enrichment value}

\item{term_col}{the ID for each gene in the format of A.B.entrez where A and
B can be of anything and entrez means entrez ID}

\item{group1}{comparison group1. Each group in \code{cluster} will be in a}

\item{group2}{comparison group2. Each group here will
be in the same tree.}

\item{simplification}{whether to simplify the GO/KEGG/Reactome terms}

\item{sim_dict}{a data frame with 2 columns: 'ori' for the original terms,
'sub' for the strings that will replace the original terms. The default is a
a limited data frame built into this package.}

\item{append_default_dict}{append default dictionary to simplify the terms
not exist.}

\item{compare_group_col_marker}{which column in \code{markers} contain the group2
information.}

\item{AP}{aesthetic parameters for plotting}

\item{...}{other parameters to pass onto \code{gg_silva} and \code{sagittae}.}

\item{separate}{tree. You may change it by specifying \code{group}=...`}
}
\description{
This function creates a compact way of visualising GSEA
results by creating multiple tree-like plots, or 'forests' (silva in Latin).
Each tree represents a pairwise comparison between one particular cell type
with all other cell types. This pattern is repeated for other cell types as
well. Each cell pair comparison is marked by an arrow (sagitta in Latin) on
top
}
