% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{ridge_all_types}
\alias{ridge_all_types}
\title{Ridge plot for multiple categories}
\usage{
ridge_all_types(
  xx,
  x_col = "value",
  y_col = "category",
  sort_by = "p.adjust",
  default_theme = F,
  color_col = "cluster",
  not_show_prop = 0.01,
  term_length = 60,
  show_num = NULL,
  simplification = T,
  AP = NULL,
  sim_dict = NULL
)
}
\arguments{
\item{xx}{a dataframe generated from run_GSEA_all_types}

\item{x_col}{the column for the x axis of ridge plot}

\item{y_col}{the column for the y axis of ridge plot}

\item{sort_by}{by which column in \code{xx} the order of the enriched terms are
shown in ridge plot}

\item{default_theme}{whether to use ggplot default_theme}

\item{color_col}{color by which column in \code{xx}}

\item{not_show_prop}{proportion of the data with extreme values of fold
changes that are not shown, in order to truncate the x axis and avoid white
spaces in the final plot.}

\item{term_length}{remove terms beyond a certain length}

\item{show_num}{how many enriched terms to show}

\item{simplification}{whether to simplify the GO/KEGG/Reactome terms}

\item{AP}{aesthetic parameters}

\item{sim_dict}{a data frame with 2 columns: 'ori' for the original terms,
'sub' for the strings that will replace the original terms. The default is a
a limited data frame built into this package.}
}
\description{
Ridge plot for multiple categories
}
\author{
Yutong Chen
}
