% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{ridge_one_type}
\alias{ridge_one_type}
\title{Ridgeplot}
\usage{
ridge_one_type(
  x,
  fill = "p.adjust",
  core_enrichment = T,
  orderBy = "NES",
  decreasing = F
)
}
\arguments{
\item{x}{an enrichment object}
}
\description{
Ridgeplot
}
\references{
\url{https://rdrr.io/github/GuangchuangYu/enrichplot/src/R/ridgeplot.R}
}
