% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{run_GSEA}
\alias{run_GSEA}
\title{GSEA}
\usage{
run_GSEA(
  markers,
  cluster_num,
  organism_db,
  organism_name = "human",
  enrich_area = "GO",
  ...
)
}
\arguments{
\item{markers}{a dataframe with at least 3 columns: one for gene names
(\code{gene_col}), one cluster (\code{cluster_col}) and one for ranking (\code{FC}). If
they are not 'feature', 'group' and 'logFC' respectively, they need to be
supplied as keyword arguments in ...}

\item{cluster_num}{which cluster to compute enrichment}

\item{organism_db}{which organism database, for example, org.Hs.eg.db}

\item{organism_name}{which organism for KEGG and Reactome}

\item{enrich_area}{either 'GO', 'KEGG' or 'reactome'}
}
\description{
GSEA
}
