% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{run_GSEA_all_types}
\alias{run_GSEA_all_types}
\title{Combine the GSEA results from all clusters}
\usage{
run_GSEA_all_types(
  markers,
  show_num = 50,
  cluster_col = "group",
  enrich_area = "KEGG",
  save_path = NULL,
  AP = NULL,
  ...
)
}
\arguments{
\item{markers}{a dataframe containing the fold changes of genes with a
column specified as \code{cluster_col} that contains the cluster information}

\item{show_num}{how many GO/KEGG terms to show. This function will select
those terms whose mean absolute fold changes are high}

\item{enrich_area}{either 'GO', 'KEGG' or 'reactome'}

\item{save_path}{save the GSEA results for later usage. GSEA is quite
computationally intensive, especially considering that it will be run for
all clusters}
}
\description{
Combine the GSEA results from all clusters
}
