% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{run_GSEA_pairwise}
\alias{run_GSEA_pairwise}
\title{Run pairwise GSEA}
\usage{
run_GSEA_pairwise(
  markers,
  organism_db,
  save_path,
  group1 = NULL,
  group2 = NULL,
  group1_col = "group",
  group2_col = "compare_group",
  ...
)
}
\arguments{
\item{markers}{a dataframe generated from \code{find_DE_genes} that contain the
log fold change of gene expression.}

\item{organism_db}{a gene database}

\item{group1}{the cell types in group1. Upregulation of genes in this group
will be reflected as positive enrichment.}

\item{group2}{the cell types in group2. Upregulation here= negative enrichment}

\item{...}{other arguments for \code{run_GSEA_all_types}. I strongly recommend
checking the correct species by specifying \code{organism_name=...}. The default
is human.}
}
\description{
Run pairwise GSEA
}
