% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_red.R
\name{RunDiffusion}
\alias{RunDiffusion}
\title{Run diffusion map}
\usage{
RunDiffusion(x, reduction = "pca", dims = 1:10)
}
\arguments{
\item{x}{Seurat object}

\item{reduction}{run diffusion map on the reduced dimension of which
method}

\item{dims}{run diffusion map on which reduced dimensions}
}
\description{
Run diffusion map
}
