% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_red.R
\name{RunGPLVM}
\alias{RunGPLVM}
\title{Add GPLVM data to Seurat dim red assay}
\usage{
RunGPLVM(x, embedding, assay = "RNA")
}
\arguments{
\item{x}{a Seurat object}

\item{embedding}{the mean pseudotime estimated from GrandPrix}
}
\description{
Add GPLVM data to Seurat dim red assay
}
