% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_multi.R
\name{sagittae}
\alias{sagittae}
\title{Draw the arrow sign for \code{rich_forest}}
\usage{
sagittae(
  plot_df,
  aes_param,
  band_ratio = 5,
  length_ratio = 1,
  nudge_x = 0.1,
  nudge_y = 1,
  yval = "yval",
  xval = "emean",
  group1 = "cluster",
  group2 = "compare_group"
)
}
\arguments{
\item{band_ratio}{how much thicker the arrow should be compared to the
\code{arrow_thickness} field in \code{AP}.}

\item{length_ratio}{the ratio of the arrow versus the maximum value along
the x axis}

\item{nudge_x}{by how much the labels for the arrow to shift away from the
arrow head.}

\item{nudge_y}{by how much the arrow and its labels to shift vertically up
from the maximum value along the y axis}
}
\description{
Please ignore the following arguments: \code{yval}, \code{xval},
\code{group1}, \code{group2}
}
