% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_gene.R
\name{save_DE_pairwise}
\alias{save_DE_pairwise}
\title{Save pairwise DE gene in excel files}
\usage{
save_DE_pairwise(DE_df, group.by, directory, ...)
}
\arguments{
\item{...}{arguments passing to \code{save_DE_genes}}
}
\description{
As many xlsx files will be created as there are comparison
groups. Each file contains the comparison groups
}
