% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_utils.R
\name{save_hub_score}
\alias{save_hub_score}
\title{Save the hub score data frame}
\usage{
save_hub_score(markers, hub_df, clusters, cell_types, save_dir)
}
\arguments{
\item{markers}{a dataframe generated from \code{find_DE_genes}}

\item{hub_df}{a dataframe generated from \code{hub_score}}
}
\description{
Save the hub score data frame
}
