% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_figures.R
\name{save_indiv_plots}
\alias{save_indiv_plots}
\title{Save a list of plots individually}
\usage{
save_indiv_plots(
  grob_list,
  save_dir,
  grid_layout,
  plot_width = 9,
  plot_height = 9,
  panel_label = NULL,
  AP = list(highlight_font = list(fontface = "bold", fontsize = 40), font_fam =
    "Arial")
)
}
\description{
This function and its arguments are based upon \code{arrange_plots}.
Instead of saving all the plots in a single pdf, this function saves the
plots individually in a directory. The usage is exactly the same as
\code{arrange_plots}.
}
