% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_point.R
\name{save_peak_genes}
\alias{save_peak_genes}
\title{Save peak genes in a particular KEGG pathway}
\usage{
save_peak_genes(
  xx,
  save_path,
  pathway = NULL,
  pathway_df = NULL,
  max_num = NULL,
  arrange_val = "val"
)
}
\description{
Save peak genes in a particular KEGG pathway
}
