% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_mat.R
\name{save_to_csv}
\alias{save_to_csv}
\title{Save Seurat object as csv}
\usage{
save_to_csv(
  x,
  directory,
  assay = NULL,
  slots = "data",
  select_cells = NULL,
  select_genes = NULL,
  label = "",
  features = c("revised_date", "revised", "date")
)
}
\arguments{
\item{x}{a seurat object}

\item{select_cells}{which cells to be saved}

\item{select_genes}{which genes to be saved, or the number of top variably
expressed genes}

\item{features}{which features in the meta data to be saved}
}
\description{
Save Seurat object as csv
}
