% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCHeat.R
\name{scaling}
\alias{scaling}
\title{Perform row or column scaling}
\usage{
scaling(xx, row_scale, column_scale)
}
\arguments{
\item{xx}{a numerical matrix or dataframe}
}
\description{
Perform row or column scaling
}
