% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ax3D.R
\name{Seg3D}
\alias{Seg3D}
\title{Extention to \code{Ax3D} with arrow heads}
\usage{
Seg3D(
  mapping = aes(group = 1),
  data = NULL,
  geom = "segment",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  AP = NULL,
  ...
)
}
\description{
Extention to \code{Ax3D} with arrow heads
}
