% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_red.R
\name{selected_cell_PCA}
\alias{selected_cell_PCA}
\title{Run PCA on selected population of cells}
\usage{
selected_cell_PCA(x, select_cells = NULL, assay = NULL, select_genes = NULL)
}
\arguments{
\item{x}{Seurat object}

\item{select_cells}{PCA on a subset of cells, a boolean vector;
default is all cells}

\item{select_genes}{PCA on a subset of genes, default is variable
genes}
}
\description{
Run PCA on selected population of cells
}
