% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCHeat_split.R
\name{seurat_heat_highlight}
\alias{seurat_heat_highlight}
\title{Highlight parts of a heatmap}
\usage{
seurat_heat_highlight(
  x,
  select_cells,
  color_row,
  group.by,
  assay = "RNA",
  slot_data = "data",
  average = F,
  return_sep = F,
  row_scale = F,
  column_scale = F,
  color_scale = NULL,
  center_scale = T,
  show_quantile = 0.05,
  AP = NULL,
  seurat_heat_params = list()
)
}
\arguments{
\item{x}{a Seurat object}

\item{select_cells}{a logical vector specifying which part of \code{x} to magnify}

\item{average}{average the expression across non highlighted cell types}

\item{return_sep}{whether the highlighted or unhighlighted heatmaps will be
returned as a list (TRUE) or as a ComplexHeatmapList (FALSE).}

\item{row_scale}{whether to perform row scaling}

\item{color_scale}{the scale of color gradient for heatmap. By default, the
\code{heatmap_color} field of the list supplied to \code{AP} should define the color
for min, middle and max of the dataset. Otherwise, please input a function
created by \code{circlize::colorRamp2}}

\item{center_scale}{whether to put white colors half way between the min and
max of the matrix. Otherwise, white color is placed at value 0.}

\item{show_quantile}{if supplied, then the color gradient would not extend
over the entire range of the matrix. The default 0.05 means that color
gradient would extend from 5\% to 95\% percentile. This is to avoid outliers
skewing the dynamic range of colors.}

\item{AP}{a list of aesthetic parameters. The important fields include
\code{cell_order} for ordering group labels, \code{gfont_fam}, \code{fontsize} and
\code{heatmap_color}}

\item{seurat_heat_params}{a named list with names being the arguments to
\code{seurat_heat}. Each item must only be a vector or list with length 1 or 2.
If the length is 2, then the first value will be for the first
(unhighlighted heatmap). The second will be for the highlighted heatmap.}
}
\value{
one or a list of ComplexHeatmaps
}
\description{
Same as \code{seurat_heat} except that a selected part of the data
is shown at a larger scale and the remainder a smaller scale
}
