% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCHeat_split.R
\name{seurat_heat_split}
\alias{seurat_heat_split}
\title{Splitting heatmap by feature}
\usage{
seurat_heat_split(
  x,
  color_row,
  group.by,
  split.by,
  main_width = NULL,
  split_color = NULL,
  ...
)
}
\description{
Splitting heatmap by feature
}
