% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPLVM_utils.R
\name{seurat_list_score}
\alias{seurat_list_score}
\title{Calculate the pathway module score for the Seurat object for each branch}
\usage{
seurat_list_score(seurat_list, save_dir, label = "DF_B", KeggID = NULL)
}
\arguments{
\item{seurat_list}{a list of seurat objects}

\item{save_dir}{where the module scores are saved}
}
\description{
Calculate the pathway module score for the Seurat object for each branch
}
