% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_gene.R
\name{seurat_violin}
\alias{seurat_violin}
\title{Similar to Seurat VlnPlot, with more graphic controls}
\usage{
seurat_violin(
  x,
  features,
  group.by,
  assay = "RNA",
  slot_data = "data",
  num_col = 3,
  AP = NULL
)
}
\description{
Similar to Seurat VlnPlot, with more graphic controls
}
