% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_gene.R
\name{seurat_volcano}
\alias{seurat_volcano}
\title{Volcano plot}
\usage{
seurat_volcano(
  markers,
  group1,
  group2,
  group1_col = "group",
  group2_col = "compare_group",
  label_genes = NULL,
  weighting = "logFC",
  pval = "padj",
  gene = "feature",
  AP = NULL
)
}
\arguments{
\item{markers}{a dataframe, result generated from \code{find_DE_genes}}

\item{group1}{comparison group1}

\item{group2}{comparison group2}

\item{group1_col}{a column in \code{markers} where \code{group1} information is stored}

\item{group2_col}{a column in \code{markers} where \code{group2} information is stored}

\item{label_genes}{which genes to show in the volcano plot}

\item{weighting}{the x axis, i.e. log fold change}

\item{pval}{the p value for the y axis}

\item{gene}{where the gene names are stored}

\item{AP}{aesthetic parameter for controlling the plot}
}
\description{
Volcano plot
}
