% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_abstract.R
\name{sigmoid}
\alias{sigmoid}
\title{Use sigmoid function to draw the curve}
\usage{
sigmoid(x, slope, off_set, elevation = 0)
}
\arguments{
\item{x}{a vector}

\item{slope}{the steepness of the curve. The larger the value, the smoother
the curve becomes}

\item{off_set}{shift the tuning point of the sigmoid curve. The higher the
value, the more leftwards the curve will shift}
}
\value{
a vector
}
\description{
Use sigmoid function to draw the curve
}
