% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_DE.R
\name{signif_p_val}
\alias{signif_p_val}
\title{Filter significant pairs of DE genes}
\usage{
signif_p_val(p_val_df)
}
\arguments{
\item{p_val_df}{output from \code{p_val_all_pairs }}
}
\value{
a matrix-like dataframe
Column names: upregulated genes in vitro1 compared with vitro2
}
\description{
Filter significant pairs of DE genes
}
