% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_emapplot.R
\name{sim_emap}
\alias{sim_emap}
\title{Plot enrichment pie chart with simplified terms}
\usage{
sim_emap(
  x,
  showCategory = 30,
  rename_vec = NULL,
  color = "p.adjust",
  layout = "nicely",
  split = NULL,
  pie = "equal",
  legend_n = 5,
  cex_category = NULL,
  pie_scale = NULL,
  cex_line = 1,
  min_edge = 0.2,
  cex_label_category = 1,
  node_label_size = NULL,
  show.legend = F,
  vert_just = 1,
  repel_text = T,
  force_repel = 1,
  AP = NULL
)
}
\arguments{
\item{split}{separate result by 'category' variable}

\item{pie}{proportion of clusters in the pie chart, one of 'equal' (default) or 'Count'}

\item{legend_n}{number of circle in legend}

\item{pie_scale}{scale of pie chart or point, this parameter has been changed to "node_scale"}

\item{cex_line}{scale of line width}

\item{min_edge}{minimum percentage of overlap genes to display the edge,
should between 0 and 1, default value is 0.2}

\item{vert_just}{vertically adjust the label position to prevent overlap
with the pie}

\item{repel_text}{whether to use ggrepel}

\item{force_repel}{how far the labels are repelled from the original
locations}
}
\description{
Plot enrichment pie chart with simplified terms
}
\author{
Guangchung Yu
}
