% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_list.R
\name{standardise_name}
\alias{standardise_name}
\title{Standardise cell labels}
\usage{
standardise_name(cellname, conversion = NULL)
}
\arguments{
\item{x}{a vector of cell names}
}
\description{
Standardise cell labels
}
