% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab3D.R
\name{Stat3D}
\alias{Stat3D}
\title{Draw 3D Geoms}
\usage{
Stat3D(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{geom}{The geom type to use \emph{ie. "point", "path", "line"}}

\item{...}{Arguements passed on to layer.}

\item{theta}{The azimuthal direction in degrees.}

\item{phi}{The colatitude in degrees.}
}
\description{
This function adds 3D geoms such as points and paths to a ggplot2 plot.
}
\references{
\url{https://github.com/AckerDWM/gg3D/blob/master/R/stat_3D.R}
}
