% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_gene.R
\name{strong_gene_violin}
\alias{strong_gene_violin}
\title{violin plot on genes that contribute to latent dimensions}
\usage{
strong_gene_violin(
  x,
  group.by,
  dim_num = 1,
  DR = "pca",
  assay = "RNA",
  data_slot = "data"
)
}
\description{
violin plot on genes that contribute to latent dimensions
}
