% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_utils.R
\name{subset_celltype}
\alias{subset_celltype}
\title{Select Seurat object by cell type}
\usage{
subset_celltype(
  x,
  top_markers = NULL,
  cluster_num = "EVT",
  select_type = "seurat_clusters"
)
}
\arguments{
\item{x}{a Seurat object}

\item{top_markers}{should be a dataframe generated by the Seurat FindMarkers
gene. Otherwise, the top variable features will be selected}

\item{cluster_num}{if it is 'all', then all cells will be selected}
}
\description{
Select Seurat object by cell type
}
