% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_terms.R
\name{subset_terms}
\alias{subset_terms}
\title{Show terms from specific groups}
\usage{
subset_terms(
  xx,
  subset_class,
  group_col = "Cluster",
  attr_name = "compareClusterResult"
)
}
\arguments{
\item{xx}{a compareClusterResult object}

\item{subset_class}{which clusters to show}
}
\description{
Show terms from specific groups
}
