% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{summarise_gsea}
\alias{summarise_gsea}
\title{Summarise the raw GSEA dataframe}
\usage{
summarise_gsea(
  xx,
  show_num,
  category_col = "category",
  cluster_col = "cluster",
  pval_col = "p.adjust",
  enrich_col = "value"
)
}
\arguments{
\item{xx}{raw dataframe from GSEA}

\item{show_num}{how many terms to show per group}
}
\value{
a shorter dataframe
}
\description{
Summarise the raw GSEA dataframe
}
