% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{term_gene_labels}
\alias{term_gene_labels}
\title{Obtain the label information for enrichment barplot}
\usage{
term_gene_labels(
  xx,
  sum_gsea,
  organism_db,
  show_gene_labels = 3,
  markers = NULL,
  category_col = "category",
  cluster_col = "cluster"
)
}
\arguments{
\item{xx}{raw dataframe from GSEA}

\item{sum_gsea}{a dataframe generated from \code{summarise_gsea}}

\item{markers}{a dataframe generated from \code{find_DE_genes}. If NULL, the
genes will be ordered alphabetically.}
}
\value{
a character vector of labels
}
\description{
This function joins the enrichment term with the top
differentially expressed genes in a single string, that will be shown in the
enrichment barplot labels.
}
