% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DR.R
\name{text_3D_repel}
\alias{text_3D_repel}
\title{3D version of \code{geom_text_repel}}
\usage{
text_3D_repel(
  dat,
  AP,
  theta,
  phi,
  label_col,
  axes_names = c("x", "y", "z"),
  repel_force = 1,
  color_text = F,
  magnify_text = 1,
  fontface = "bold",
  ...
)
}
\arguments{
\item{dat}{labelling data frame}

\item{AP}{aesthetic parameters, only need to supply \code{point_fontsize}}

\item{theta}{angle of azimuthal rotation}

\item{phi}{angle of vertical rotation}

\item{axes_names}{the column names for the x, y, and z coordinates in \code{dat}}

\item{repel_force}{extent of repelling labels}

\item{color_text}{whether to color the label text according to \code{label_col}}

\item{magnify_text}{how much bigger to make the label text}
}
\value{
a \code{geom_text_repel} layer
}
\description{
Same usage as \code{geom_text_repel} by adding it after a ggplot
object
}
