% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TF}
\alias{TF}
\title{A list of transcriptional factors}
\format{
a vector
}
\usage{
TF
}
\description{
A list of transcriptional factors
}
\references{
\url{https://www.cell.com/cell/comments/S0092-8674(18)30106-5#supplementaryMaterial}
\url{http://humantfs.ccbr.utoronto.ca/download.php}
}
\keyword{datasets}
