% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_motif.R
\name{TF_enrichment_pairwise}
\alias{TF_enrichment_pairwise}
\title{Compare TF motif enrichment between cell types}
\usage{
TF_enrichment_pairwise(markers, compare_list, motif_dir, save_path, ...)
}
\arguments{
\item{motif_dir}{where the TF motifs are stored}

\item{...}{refer to \code{DE_gene_pairwise_list}}

\item{save_dir}{save the results of the analysis which is computationally
intensive}
}
\description{
Please ensure 'RcisTarget' is installed
}
