% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{theme_dim_red}
\alias{theme_dim_red}
\title{Theme for DimPlot}
\usage{
theme_dim_red(aes_param, color_fill)
}
\description{
For dimensionality reduction plots, use a minimalistic theme
without any borders, grids or axis.
}
\examples{
ggplot (x, aes (x, y) ) +
geom_point () +
theme_dim_red ()
}
