% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{theme_TB}
\alias{theme_TB}
\title{Append this function at the end of a ggplot to provide the style in this
paper}
\usage{
theme_TB(
  plot_type = "no_arrow",
  plot_ob = NULL,
  feature_vec = NULL,
  color_fill = F,
  color_vec = NULL,
  rotation = 90,
  more_prec = 2,
  aes_param = list(fontsize = 15, point_fontsize = 6, font_fam = "Arial", pointsize =
    3, legend_point_size = 5),
  ...
)
}
\arguments{
\item{plot_type}{'dim_red' or 'dotplot'. In 'dim_red', no grid lines will be
shown and arrow signs are added while in 'dotplot', grid lines will be shown}

\item{plot_ob}{if 'dim_red' is selected in \code{plot_type}, then it is necessary
to supply the ggplot object. This is to extract the axis information using
\code{ggplot_build}, which is then used to draw the axis arrows. I am attempting
to create a new geom layer but so far \code{ggproto} does not seem to allow me to
extract such information.
NB: if automated coloring is desired, it is also essential to input plot_ob}

\item{color_fill}{whether the color schemes will be applied to
\code{scale_fill_continuous}}

\item{rotation}{by how much degree the x axis text is rotated}

\item{aes_param}{a list of parameters controlling plot aesthetics}

\item{...}{keyword argument to \code{arrow_axis}}

\item{whether}{to rotate the x axis labels or not}

\item{length_ratio}{the length of the arrows for dim_plot as a percentage of
the plot size}
}
\value{
a list of ggplot layers that can be concatenated to a ggplot object
}
\description{
Append this function at the end of a ggplot to provide the style in this
paper
}
