% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPLVM_utils.R
\name{time_cluster_plot}
\alias{time_cluster_plot}
\title{Plot the clustering results of times against the peak gradient}
\usage{
time_cluster_plot(
  peak_plot,
  metadata,
  show_text_prop = 0.95,
  color_by = "cluster",
  color_bar = "broad_type",
  time_col = "MGP_PT",
  exclude_perc = 0.02,
  vjust = 0,
  thickness_ratio = 0.05,
  repel_force = 1,
  repel_point = NULL,
  ribbon_alpha = 1,
  AP = NULL,
  overlap = T
)
}
\arguments{
\item{peak_plot}{a dataframe generated by \code{find_peak} or
\code{find_transition_from_seurat}. It must contain the column \code{peak_time} and \code{val},
the latter for selection of the most significant genes}

\item{metadata}{a dataframe for adding the cell types colors}

\item{color_by}{which column in \code{peak_plot} to color the points}

\item{color_bar}{which column in \code{metadata} contains cell type information
to provide a cellular context for the pseudotime line}

\item{time_col}{which column in \code{metadata} contains pseudotime information
for each cell}

\item{exclude_perc}{by how much percent the pseudotime line should be
truncated. This is because pseudotime points sometimes contain some outliers
that may create a lot of white space.}
}
\description{
Plot the clustering results of times against the peak gradient
}
