% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_utils.R
\name{WGCNA_heatmap}
\alias{WGCNA_heatmap}
\title{Plot heatmap for WGCNA gene cluster}
\usage{
WGCNA_heatmap(
  x,
  gene_names,
  celltype,
  feature,
  main_width = 25,
  side_width = 5
)
}
\description{
This function is deprecated. Please use \code{seurat_heat}
}
