% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNA_utils.R
\name{write_module_gene}
\alias{write_module_gene}
\title{Record genes of a particular module}
\usage{
write_module_gene(datExpr, filename, net, save_dir, top_marks = NULL)
}
\description{
Record genes of a particular module
}
