% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{add_custom_color}
\alias{add_custom_color}
\title{Customise the colors that correspond to particular features}
\usage{
add_custom_color(feature_vec, aes_param, color_fill = F)
}
\arguments{
\item{feature_vec}{a vector of features for which colors will be assigned}

\item{color_fill}{whether scale_color_* or scale_fill_* is used}
}
\description{
Customise the colors that correspond to particular features
}
