% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ax2D.R
\name{arrow_axis}
\alias{arrow_axis}
\title{Create miniature arrow to replace axes}
\usage{
arrow_axis(
  plot_ob,
  length_ratio = 0.05,
  nudge_ratio = 0,
  move_x = 0,
  move_y = 0,
  reverse_x = F,
  reverse_y = F,
  axis_label_sim = F,
  nudge_dimname = 0,
  aes_param = list(pointsize = 3, font_fam = "Arial", point_fontsize = 6)
)
}
\arguments{
\item{length_ratio}{the ratio of the length of the arrow relative to the
span of the data}

\item{nudge_ratio}{how much the arrow labels should move away from the arrow
tip as a ratio of the length of the arrow}

\item{move_x}{move the arrow axies position by how much to the left,
negative value is to the right}

\item{axis_label_sim}{whether to use a simple label i.e. 'dim1', 'dim2'}

\item{nudge_dimname}{if \code{axis_label_sim} is True, then how much further to
the right to move the dim name label}

\item{aes_param}{setting for pointsize, font_fam, point_fontsize,
arrow_angle, arrow_length, arrow_length_unit, arrow_type, arrow_thickness,
arrow_linejoin}

\item{return}{a list of geoms for the arrow segment and axis labels}
}
\description{
Create miniature arrow to replace axes
}
