% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_list.R
\name{clean_metadata}
\alias{clean_metadata}
\title{Clean metadata of Seurat objects}
\usage{
clean_metadata(x)
}
\description{
The cell type labels are converted into levelled factors. 'D'
is appended before data labels.
}
