% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overRep.R
\name{compare_cluster_enrichment}
\alias{compare_cluster_enrichment}
\title{Over-representation test of genes in each cluster}
\usage{
compare_cluster_enrichment(
  markers,
  GO_data,
  organism_db,
  organism_name = "human",
  logFC_term = "logFC",
  pval_term = "padj",
  gene_name = "feature",
  log_FC_thres = 0.25,
  pval_thres = 0.05,
  enrich_area = "KEGG"
)
}
\arguments{
\item{markers}{a dataframe of DE genes from Seurat; alternatively, a list of
DE genes may be used}

\item{GO_data}{result from go_data}

\item{organism_db}{which organism database, for example, org.Hs.eg.db}

\item{organism_name}{which organism for KEGG and Reactome, e.g. 'human'. If
\code{enrich_area} is TF or MIRNA, the organism name should be for the gprofiler
database, e.g.  'hsapiens', 'mmusculus'}

\item{logFC_term}{which column in \code{markers} contains the information for
ranking gene expression}

\item{enrich_area}{either KEGG or GO or reactome}

\item{logFC_thres}{threshold of \code{logFC_term} beyond which a gene is selected
for over-representation test}
}
\description{
Over-representation test of genes in each cluster
}
