% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{compute_all_cor}
\alias{compute_all_cor}
\title{Compute the difference between every single pair of cells}
\usage{
compute_all_cor(
  x,
  method = "correlation",
  assay = "integrated",
  select_cells1 = NULL,
  select_cells2 = NULL,
  select_genes = NULL
)
}
\arguments{
\item{x}{a Seurat object}

\item{method}{'correlation', 'partial_corr', or 'distance'}

\item{select_cells1}{a logical vector selecting the cells forming the rows
of the correlation matrix}

\item{select_cells2}{similarly for the coluns}

\item{select_genes}{which genes to compare, default is variable genes}
}
\description{
Compute the distance matrix using a user-defined metric
}
