% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_gene.R
\name{find_DE_genes}
\alias{find_DE_genes}
\title{Find DE genes across a particular feature}
\usage{
find_DE_genes(
  x,
  directory,
  feature = "seurat_clusters",
  filename = NULL,
  load_data = TRUE,
  label = "all"
)
}
\arguments{
\item{x}{a Seurat object}

\item{directory}{where to save the computed markers}

\item{feature}{a column in the meta data}

\item{load_data}{whether to load existing marker data}
}
\description{
Find DE genes across a particular feature
}
