% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_preprocess.R
\name{frequency_plot}
\alias{frequency_plot}
\title{Boxplot or Violin plot the count of cells expressing a certain marker}
\usage{
frequency_plot(
  x_df,
  gene,
  group.by = "condition",
  rep.by = "series",
  box_plot = T,
  express_thres = 7,
  perc = T
)
}
\arguments{
\item{x_df}{a dataframe}

\item{gene}{which gene to plot}

\item{group.by}{the experimental conditions}

\item{rep.by}{the replication for each condition}

\item{box_plot}{whether to use boxplot or violin plot}

\item{express_thres}{the fluorescence value by which a cell is called
positive for a gene}

\item{perc}{whether to show the percentage of positive cells or the number
of positive cells}
}
\description{
Boxplot or Violin plot the count of cells expressing a certain marker
}
