% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_gene.R
\name{gene_gene_mat}
\alias{gene_gene_mat}
\title{Plot the gene}
\usage{
gene_gene_mat(
  x,
  compare.by,
  compare_items,
  group.by,
  assay = "RNA",
  data_slot = "data"
)
}
\arguments{
\item{x}{a Seurat object}

\item{compare.by}{which features to be compared}

\item{compare_items}{which 2 features in \code{compare.by} to be compared}

\item{group.by}{perform the analysis separately for different groups of
cells}
}
\description{
Plot the gene
}
