% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPLVM_utils.R
\name{gene_over_pseudotime}
\alias{gene_over_pseudotime}
\title{Gene expression with uncertainty over pseudotime}
\usage{
gene_over_pseudotime(
  x,
  exp_mat,
  genes,
  metadata,
  color_feature,
  num_col = 4,
  num_row = NULL,
  branch_assignment = NULL,
  peak_data = NULL,
  gene_col = "feature",
  AP = NULL
)
}
\arguments{
\item{x}{dataframe containing mean + var information estimated by BRGP}

\item{exp_mat}{the expression matrix, i.e. true expression levels}

\item{genes}{the genes to be plotted}
}
\description{
Gene expression with uncertainty over pseudotime
}
