% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_DR.R
\name{gg_DimPlot}
\alias{gg_DimPlot}
\title{Reimplementation of DimPlot in Seurat for better graphic controls}
\usage{
gg_DimPlot(
  x,
  feature,
  DR = "pca",
  dims = c(1, 2),
  size_highlight = NULL,
  highlight_font = 4,
  default_color = F,
  further_repel = T,
  AP = NULL,
  ...
)
}
\arguments{
\item{x}{a Seurat object}

\item{size_highlight}{a character, logical or numeric vector specifying
which cells to magnify in size}

\item{highlight_font}{how large the highlighted cells should be}

\item{further_repel}{if TRUE, the labels would be repelled away from the
data points as much as possible}

\item{AP}{a list for \code{custom_color} and \code{theme_TB}}
}
\description{
Reimplementation of DimPlot in Seurat for better graphic controls
}
\author{
Yutong Chen
}
