% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_analysis.R
\name{make_cycle_heat}
\alias{make_cycle_heat}
\title{Make a heatmap for phase score}
\usage{
make_cycle_heat(
  x,
  seurat_ob,
  group.by = NULL,
  AP = NULL,
  direction = "vertical",
  ...
)
}
\arguments{
\item{x}{a matrix containing the phase scores}

\item{seurat_ob}{a seurat object that provides the metadata}

\item{group.by}{which feature to plot the row color bar for the heatmap.
Only 1 feature allowed if \code{direction} is vertical}

\item{direction}{either vertical or horizontal}
}
\description{
Make a heatmap for phase score
}
