% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_abstract.R
\name{plot_two_branch}
\alias{plot_two_branch}
\title{Draw 2 sigmoid curves}
\usage{
plot_two_branch(
  dat,
  branch_col = "epil_branch",
  color_col = "broad_type",
  time_col = "MGP_PT",
  branch1 = c("TB_stem", "STB_branch"),
  branch2 = c("TB_stem", "EVT_branch"),
  select_cells1 = NULL,
  select_cells2 = NULL,
  AP = NULL,
  time_axis = T,
  axis_label = "time",
  vert_trans = 0,
  branch_names = c(NA, NA, NA),
  plot_branch_names = T,
  shorten_ratio = 0.9,
  thickness = 0.01,
  show_tick = T,
  ...
)
}
\arguments{
\item{branch1}{which samples in \code{dat} make up the first branch}

\item{select_cells1}{which cells in the first branch}

\item{time_axis}{whether to plot the x axis}

\item{axis_label}{what to label the \code{time_axis}}

\item{vert_trans}{the extent of vertical translation}
}
\value{
a list of ggplot layers
}
\description{
Draw 2 sigmoid curves
}
\examples{
ggplot () +
plot_two_branch (meta, 
                 select_cells1 = meta$broad_type != 'STB',
                 select_cells2 = meta$broad_type != 'EVT',
                 text_offset=0.007, off_set= 5
) + ggplot2::theme (legend.position='none')
}
\seealso{
\code{plot_one_branch}
}
