% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPLVM_utils.R
\name{raw_to_seurat}
\alias{raw_to_seurat}
\title{Obtain the mean values and store them in Seurat objects according to the
branch assignment indicated by \code{label_list}}
\usage{
raw_to_seurat(pred_all, label_list)
}
\description{
Obtain the mean values and store them in Seurat objects according to the
branch assignment indicated by \code{label_list}
}
