% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_red.R
\name{run_dim_red}
\alias{run_dim_red}
\title{Run DR}
\usage{
run_dim_red(
  x,
  assay = NULL,
  pca_features = NULL,
  select_cells = NULL,
  feature_num = 2000,
  run_diff_map = FALSE,
  save_mem = NULL,
  normalize = FALSE,
  run_umap = TRUE,
  var_scale = FALSE,
  cluster = TRUE,
  find_var_features = FALSE
)
}
\arguments{
\item{assay}{assay to use for dimensionality reduction}

\item{pca_features}{run PCA on a selected list of genes, should be a
character vector}

\item{select_cells}{run PCA on a selected population of cells;
should be a boolean vector}

\item{run_diff_map}{whether to run diffusion map using the package
destiny. Depending on the dataset, this may take several minutes}

\item{save_mem}{for large matrices whether the entire matrix should be
stored, or simply storing the \code{VariableFeatures}}

\item{normalize}{whether to normalize the data}

\item{cluster}{whether to perform clustering as well}

\item{scale}{the variable features only to reduce memory}
}
\description{
Run DR
}
