% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_gene.R
\name{save_DE_genes}
\alias{save_DE_genes}
\title{Save Differentially Expressed Genes}
\usage{
save_DE_genes(
  x,
  save_dir,
  group.by = "revised",
  label = "all",
  show_num = 30,
  weight = "logFC",
  cluster = "group"
)
}
\arguments{
\item{x}{a Seurat object}
}
\description{
Save Differentially Expressed Genes
}
