% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_mat.R
\name{save_to_csv}
\alias{save_to_csv}
\title{Save Seurat object as csv}
\usage{
save_to_csv(
  x,
  directory = NULL,
  assay = NULL,
  slot_data = "data",
  select_cells = NULL,
  select_genes = NULL,
  label = "",
  save_result = T
)
}
\arguments{
\item{x}{a seurat object}

\item{directory}{where the results are saved}

\item{select_cells}{which cells to be saved}

\item{select_genes}{which genes to be saved, or the number of top variably
expressed genes}

\item{save_result}{if FALSE, the expression matrix is the output. If no
directory is supplied, this argument is set FALSE automatically}
}
\description{
save the expression matrix and metadata of a Seurat object
}
