% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_gene.R
\name{seurat_violin}
\alias{seurat_violin}
\title{Similar to Seurat VlnPlot, with more graphic controls}
\usage{
seurat_violin(
  x,
  features,
  group.by,
  assay = "RNA",
  slot_data = "data",
  num_col = NULL,
  free_xy = "fixed",
  AP = NULL,
  plot_type = "box",
  color.by = NULL,
  ...
)
}
\arguments{
\item{x}{a Seurat object}

\item{features}{genes to plot, which will be split over by facets}

\item{group.by}{by which cell feature}

\item{num_col}{in how many columns the facets are arranged}

\item{plot_type}{can be 'box', 'violin' or 'scatter'.}

\item{...}{additional arguments for \code{custom_tick}}

\item{color_by}{only effective when plot_type is 'scatter'}
}
\description{
This function was called 'seurat_violin' originally because of
re-implementation of violin plot in seurat. Its functionality has now
expanded to include scatter plots and boxplots.
}
