% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_gene.R
\name{seurat_violin}
\alias{seurat_violin}
\title{Similar to Seurat VlnPlot, with more graphic controls}
\usage{
seurat_violin(
  x,
  features,
  group.by,
  assay = "RNA",
  slot_data = "data",
  num_col = NULL,
  free_xy = "fixed",
  AP = NULL,
  box_plot = T,
  ...
)
}
\arguments{
\item{x}{a Seurat object}

\item{features}{genes to plot, which will be split over by facets}

\item{group.by}{by which cell feature}

\item{num_col}{in how many columns the facets are arranged}

\item{...}{additional arguments for \code{custom_tick}}
}
\description{
Similar to Seurat VlnPlot, with more graphic controls
}
