% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_gene.R
\name{seurat_volcano}
\alias{seurat_volcano}
\title{Deseq2 style Volcano plot}
\usage{
seurat_volcano(
  markers,
  group1,
  group2,
  group1_col = "group",
  group2_col = "compare_group",
  label_genes = NULL,
  weighting = "logFC",
  pval = "padj",
  gene = "feature",
  weight_thres = 0.25,
  show_gene_num = 30,
  AP = NULL,
  ...
)
}
\arguments{
\item{markers}{a dataframe, result generated from \code{find_DE_genes}}

\item{group1}{comparison group1}

\item{group2}{comparison group2}

\item{group1_col}{a column in \code{markers} where \code{group1} information is stored}

\item{group2_col}{a column in \code{markers} where \code{group2} information is stored}

\item{label_genes}{which genes to show in the volcano plot}

\item{weighting}{the x axis, i.e. log fold change}

\item{pval}{the p value for the y axis}

\item{gene}{where the gene names are stored}

\item{weight_thres}{above which value in \code{weighting} are the points colored
according to their comparison groups}

\item{AP}{aesthetic parameter for controlling the plot}

\item{...}{arguments for \code{enrich_arrow}, including \code{nudge_x}, \code{nudge_y} and
\code{band_thickness}, \code{length_ratio}. For description of their roles, see
\code{enrich_bar}}
}
\value{
a ggplot object
}
\description{
Deseq2 style Volcano plot
}
