% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_list.R
\name{standardise_name}
\alias{standardise_name}
\title{Standardise cell labels}
\usage{
standardise_name(cellname, conversion = NULL)
}
\arguments{
\item{cellname}{a vector of cell names}

\item{conversion}{the dataframe to convert cell names}
}
\value{
a character vector based on \code{cellname}
}
\description{
Standardise cell labels
}
