% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_gene.R
\name{unique_DE_genes}
\alias{unique_DE_genes}
\title{Select the top DE genes from each cluster
If 1 DE gene appears in 2 clusters, the cluster with the highest avg_logFC
will be chosen
NB: only upregulated genes will be selected}
\usage{
unique_DE_genes(
  DE_genes,
  top_number,
  gene = "feature",
  weighting = "logFC",
  cluster = "group"
)
}
\arguments{
\item{DE_genes}{a dataframe}

\item{top_number}{how many DE genes per cluster/group to choose}

\item{gene}{which column has the gene symbols}

\item{weighting}{which column has the gene rankings}

\item{cluster}{which column has the cluster names/numbers}
}
\description{
Select the top DE genes from each cluster
If 1 DE gene appears in 2 clusters, the cluster with the highest avg_logFC
will be chosen
NB: only upregulated genes will be selected
}
